### 2022-09-27 6a1bc65
model.py: support scalar inputs in add_constraints
