+++ doc/release_notes.rst
@@ -5,7 +5,7 @@ Upcoming Release
 ----------------
 
 * AnonymousConstraint can now built from comparison operations of variables with constants, e.g. `x >= 5`.
-
+* The `Model.add_constraints` function now support input of type `ScalarVariable`, `ScalarLinearExpression` and `ScalarConstraint`.
 
 Version 0.0.11
 --------------
