### 2022-09-28 c4c3949
* add ipython completion * fix inplace update of constraints' LHS
