### 2022-09-30 e87cbdf
feat: filter out zeros before writing to file, add sanitize_zeros function
