### 2022-10-06 #4823 @editorialbot
[REVIEW]: Linopy: Linear optimization with n-dimensional labeled variables
**Submitting author:** @fabianhofmann (<a href="http://orcid.org/0000-0002-6604-5450">Fabian Hofmann</a>)
**Repository:** https://github.com/PyPSA/linopy
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.11
