### 2022-10-10 cef5200
add matrices module and matrix accessor (#44)
