### 2022-10-10 61e2829
io: use matrices accessor for directly exporting to gurobipy
