### 2022-10-10 311e9a3
matrices: fix doc strings
