### 2022-10-10 47e7a30
Use matrices accessor for directly exporting to gurobipy (#45)
* io: use matrices accessor for directly exporting to gurobipy

* matrices: fix doc strings
