+++ test/test_matrices.py
@@ -7,6 +7,7 @@ Created on Mon Oct 10 14:21:23 2022.
 """
 
 import numpy as np
+import pandas as pd
 import xarray as xr
 
 from linopy import Model
@@ -24,6 +25,26 @@ def test_basic_matrices():
 
     obj = (10 * x + 5 * y).sum()
     m.add_objective(obj)
+
+    assert m.matrices.A.shape == (*m.matrices.clabels.shape, *m.matrices.vlabels.shape)
+    assert m.matrices.clabels.shape == m.matrices.sense.shape
+    assert m.matrices.vlabels.shape == m.matrices.ub.shape
+    assert m.matrices.vlabels.shape == m.matrices.lb.shape
+
+
+def test_basic_matrices_masked():
+    m = Model()
+
+    lower = pd.Series(0, range(10))
+    x = m.add_variables(lower, name="x")
+    mask = pd.Series([True] * 8 + [False, False])
+    y = m.add_variables(lower, name="y", mask=mask)
+
+    m.add_constraints(x + y, ">=", 10)
+
+    m.add_constraints(y, ">=", 0)
+
+    m.add_objective(2 * x + y)
 
     assert m.matrices.A.shape == (*m.matrices.clabels.shape, *m.matrices.vlabels.shape)
     assert m.matrices.clabels.shape == m.matrices.sense.shape
