### 2022-10-11 7f66a5e
io: remove hard drop zeros, make it optional instead
