### 2022-10-11 22a9665
pips: add combined row labels for submatrices beside each other
