### 2022-10-12 0acb4d6
expressions: 	* allow ScalarVariable as input for LinearExpression.from_tuples 	* allow pandas.Series as input for groupby_sum
