### 2022-10-12 2d89a51
io: add to_highspy function
