### 2022-10-13 282b822
use highspy instead of shell commands, enable direct io api
