+++ test/test_optimization.py
@@ -17,6 +17,7 @@ from linopy.solvers import available_solvers
 params = [(name, "lp") for name in available_solvers]
 if "gurobi" in available_solvers:
     params.append(("gurobi", "direct"))
+    params.append(("highs", "direct"))
 
 
 @pytest.fixture
