+++ .github/workflows/CI.yaml
@@ -47,7 +47,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -e .[dev]
-        pip install cplex xpress gurobipy
 
     - name: Lint with flake8
       run: |
