### 2022-10-13 39137f2
ci: only install highspy for linux
