+++ .github/workflows/CI.yaml
@@ -26,6 +26,7 @@ jobs:
       run: |
         sudo apt-get install glpk-utils
         sudo apt-get install coinor-cbc
+        pip install highspy
 
     - name: Install macos dependencies
       if: matrix.os == 'macos-latest'
