### 2022-10-13 46f168c
test: fix available solver selection
