### 2022-10-13 61871b4
setup.py: revert inclusion of highspy
