### 2022-10-13 1bfdb27
highs: print out current log file
