### 2022-10-13 70a9a8a
Integrate highspy (#48)
* io: add to_highspy function

* use highspy instead of shell commands, enable direct io api

* update release notes
update setup.py

* ci: only install highspy for linux

* test: fix available solver selection

* setup.py: revert inclusion of highspy

* highs: print out current log file
