### 2022-10-14 3b8a65e
doc: minor fix in API reference (#49)
Line was duplicated.
