### 2022-10-14 a7f472b
pips: add combined row labels for submatrices beside each other (#50)
