### 2022-10-17 50e4cd3
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/codespell-project/codespell: v2.2.1 → v2.2.2](https://github.com/codespell-project/codespell/compare/v2.2.1...v2.2.2)
