### 2022-11-03 e882781
Refactor LinearExpression and Constraint using composition
Since there are unsafe methods in xr.Dataset that do not correctly
pass on the LinearExpression or Constraint behaviour, we switch
the two classes to a composition instead of inheritance scheme.
