+++ test/test_linear_expression.py
@@ -14,6 +14,7 @@ from xarray.testing import assert_equal
 
 from linopy import LinearExpression, Model, merge
 from linopy.expressions import ScalarLinearExpression
+from linopy.testing import assert_linequal
 
 m = Model()
 
@@ -43,7 +44,7 @@ def test_values():
 
 def test_duplicated_index():
     expr = m.linexpr((10, "x"), (-1, "x"))
-    assert (expr._term == [0, 1]).all()
+    assert (expr.data._term == [0, 1]).all()
 
 
 def test_variable_to_linexpr():
@@ -57,23 +58,23 @@ def test_variable_to_linexpr():
 
     expr = 10 * x + y
     assert isinstance(expr, LinearExpression)
-    assert_equal(expr, m.linexpr((10, "x"), (1, "y")))
+    assert_linequal(expr, m.linexpr((10, "x"), (1, "y")))
 
     expr = x + 8 * y
     assert isinstance(expr, LinearExpression)
-    assert_equal(expr, m.linexpr((1, "x"), (8, "y")))
+    assert_linequal(expr, m.linexpr((1, "x"), (8, "y")))
 
     expr = x + y
     assert isinstance(expr, LinearExpression)
-    assert_equal(expr, m.linexpr((1, "x"), (1, "y")))
+    assert_linequal(expr, m.linexpr((1, "x"), (1, "y")))
 
     expr = x - y
     assert isinstance(expr, LinearExpression)
-    assert_equal(expr, m.linexpr((1, "x"), (-1, "y")))
+    assert_linequal(expr, m.linexpr((1, "x"), (-1, "y")))
 
     expr = -x - 8 * y
     assert isinstance(expr, LinearExpression)
-    assert_equal(expr, m.linexpr((-1, "x"), (-8, "y")))
+    assert_linequal(expr, m.linexpr((-1, "x"), (-8, "y")))
 
     expr = np.array([1, 2]) * x
     assert isinstance(expr, LinearExpression)
@@ -155,7 +156,7 @@ def test_add():
     assert res.nterm == expr.nterm + other.nterm
     assert (res.coords["dim_0"] == expr.coords["dim_0"]).all()
     assert (res.coords["dim_1"] == other.coords["dim_1"]).all()
-    assert res.notnull().all().to_array().all()
+    assert res.data.notnull().all().to_array().all()
 
     assert isinstance(x - expr, LinearExpression)
     assert isinstance(x + expr, LinearExpression)
@@ -174,7 +175,7 @@ def test_sub():
     assert res.nterm == expr.nterm + other.nterm
     assert (res.coords["dim_0"] == expr.coords["dim_0"]).all()
     assert (res.coords["dim_1"] == other.coords["dim_1"]).all()
-    assert res.notnull().all().to_array().all()
+    assert res.data.notnull().all().to_array().all()
 
 
 def test_sum():
@@ -182,14 +183,14 @@ def test_sum():
     res = expr.sum("dim_0")
 
     assert res.size == expr.size
-    assert res.nterm == expr.nterm * len(expr.dim_0)
+    assert res.nterm == expr.nterm * len(expr.data.dim_0)
 
     res = expr.sum()
     assert res.size == expr.size
     assert res.nterm == expr.size
-    assert res.notnull().all().to_array().all()
+    assert res.data.notnull().all().to_array().all()
 
-    assert_equal(expr.sum(["dim_0", "_term"]), expr.sum("dim_0"))
+    assert_linequal(expr.sum(["dim_0", "_term"]), expr.sum("dim_0"))
 
 
 def test_where():
@@ -209,10 +210,10 @@ def test_merge():
     expr2 = z.sum("dim_0")
 
     res = merge(expr1, expr2)
[…]
