+++ linopy/constraints.py
@@ -6,6 +6,7 @@ This module contains implementations for the Constraint{s} class.
 """
 
 import re
+from deprecated import deprecated
 from dataclasses import dataclass
 from itertools import product
 from typing import Any, Sequence, Union
@@ -21,51 +22,29 @@ from xarray import DataArray, Dataset
 from linopy import expressions, variables
 from linopy.common import (
     _merge_inplace,
-    has_assigned_model,
     has_optimized_model,
     is_constant,
     replace_by_map,
 )
 
 
-class Constraint(DataArray):
+@dataclass(repr=False)
+class Constraint:
     """
-    Constraint container for storing constraint labels.
+    Projection to a single constraint in a model
 
     The Constraint class is a subclass of xr.DataArray hence most xarray
     functions can be applied to it.
     """
 
-    __slots__ = ("_cache", "_coords", "_indexes", "_name", "_variable", "model")
-
-    def __init__(self, *args, **kwargs):
-
-        # workaround until https://github.com/pydata/xarray/pull/5984 is merged
-        if isinstance(args[0], DataArray):
-            da = args[0]
-            args = (da.data, da.coords)
-            kwargs.update({"attrs": da.attrs, "name": da.name})
-
-        self.model = kwargs.pop("model", None)
-        super().__init__(*args, **kwargs)
-        assert self.name is not None, "Constraint data does not have a name."
-
-    # We have to set the _reduce_method to None, in order to overwrite basic
-    # reduction functions as `sum`. There might be a better solution (?).
-    _reduce_method = None
-
-    # Disable array function, only function defined below are supported
-    # and set priority higher than pandas/xarray/numpy
-    __array_ufunc__ = None
-    __array_priority__ = 10000
+    name: str
+    model: "Model"
 
     def __repr__(self):
         """
         Get the string representation of the constraints.
         """
-        data_string = (
-            "Constraint labels:\n" + self.to_array().__repr__().split("\n", 1)[1]
-        )
+        data_string = "Constraint labels:\n" + self.labels.__repr__().split("\n", 1)[1]
         extend_line = "-" * len(self.name)
         return (
             f"Constraint '{self.name}':\n"
@@ -78,18 +57,28 @@ class Constraint(DataArray):
         Get the html representation of the variables.
         """
         # return self.__repr__()
-        data_string = self.to_array()._repr_html_()
+        data_string = self.labels._repr_html_()
         data_string = data_string.replace("xarray.DataArray", "linopy.Constraint")
         return data_string
 
+    @deprecated(
+        reason="Constraint.to_array has been replaced by using the .labels property"
+    )
     def to_array(self):
         """
         Convert the variable array to a xarray.DataArray.
         """
-        return DataArray(self)
+        return self.labels
 
     @property
-    @has_assigned_model
+    def labels(self):
+        return self.model.constraints.labels[self.name]
+
+    @labels.setter
+    def labels(self, value):
+        raise RuntimeError("Labels are read-only")
+
+    @property
     def coeffs(self):
         """
         Get the left-hand-side coefficients of the constraint.
@@ -100,7 +89,6 @@ class Constraint(DataArray):
         return self.model.constraints.coeffs[self.name]
 
     @coeffs.setter
-    @has_assigned_model
     def coeffs(self, value):
         term_dim = self.name + "_term"
         value = DataArray(value).broadcast_like(self.vars, exclude=[term_dim])
@@ -115,7 +103,6 @@ class Constraint(DataArray):
             self.model.constraints.coeffs = coeffs
 
     @property
-    @has_assigned_model
     def vars(self):
[…]

[…]
