### 2022-11-03 ef9d2ee
Small fixes introduced during clean-up
