+++ linopy/testing.py
@@ -3,4 +3,4 @@ from .expressions import _expr_unwrap
 
 
 def assert_linequal(a, b):
-    return assert_equal(_expr_unrwap(a), _expr_unwrap(b))
+    return assert_equal(_expr_unwrap(a), _expr_unwrap(b))
