### 2022-11-03 d94214d
Use deprecation instead of deprecated package
