+++ linopy/constraints.py
@@ -6,16 +6,15 @@ This module contains implementations for the Constraint{s} class.
 """
 
 import re
-from deprecation import deprecated
 from dataclasses import dataclass
 from itertools import product
 from typing import Any, Sequence, Union
 
+from deprecation import deprecated
 import dask
 import numpy as np
 import pandas as pd
 import xarray as xr
-from deprecated import deprecated
 from numpy import arange, array
 from scipy.sparse import coo_matrix
 from xarray import DataArray, Dataset
