+++ linopy/expressions.py
@@ -11,8 +11,8 @@ import logging
 from dataclasses import dataclass
 from itertools import product, zip_longest
 from typing import Union
-from deprecation import deprecated
 
+from deprecation import deprecated
 import numpy as np
 import pandas as pd
 import xarray as xr
