### 2022-11-07 e10c096
variables: Convert into composition pattern (WIP)
