+++ test/test_io.py
@@ -41,7 +41,7 @@ def test_str_arrays_with_nans():
     x = m.add_variables(4, pd.Series([8, 10]), name="x")
     # now expand the second dimension, expended values of x will be nan
     y = m.add_variables(0, pd.DataFrame([[1, 2], [3, 4], [5, 6]]), name="y")
-    assert m["x"].data[-1].item() == -1
+    assert m["x"].values[-1] == -1
 
     da = int_to_str(m["x"].values)
     assert da.dtype == object
