+++ linopy/variables.py
@@ -15,6 +15,7 @@ from warnings import warn
 import dask
 import numpy as np
 import pandas as pd
+from deprecation import deprecated
 from numpy import floating, inf, issubdtype
 from xarray import DataArray, Dataset, zeros_like
 
@@ -24,16 +25,15 @@ from linopy.common import (
     has_assigned_model,
     has_optimized_model,
     is_constant,
+    forward_as_properties,
 )
-
 
 def varwrap(method, *default_args, **new_default_kwargs):
     @functools.wraps(method)
-    def _varwrap(obj, *args, **kwargs):
+    def _varwrap(var, *args, **kwargs):
         for k, v in new_default_kwargs.items():
             kwargs.setdefault(k, v)
-        obj = DataArray(obj)
-        return Variable(method(obj, *default_args, *args, **kwargs))
+        return var.__class__(method(var.labels, *default_args, *args, **kwargs))
 
     _varwrap.__doc__ = f"Wrapper for the xarray {method} function for linopy.Variable"
     if new_default_kwargs:
@@ -42,7 +42,19 @@ def varwrap(method, *default_args, **new_default_kwargs):
     return _varwrap
 
 
-class Variable(DataArray):
+def _var_unwrap(var):
+    if isinstance(var, Variable):
+        return var.labels
+    return var
+
+
+@dataclass(repr=False)
+@forward_as_properties(
+    labels=[
+        "name"
+    ]
+)
+class Variable:
     """
     Variable container for storing variable labels.
 
@@ -91,28 +103,14 @@ class Variable(DataArray):
     Further operations like taking the negative and subtracting are supported.
     """
 
-    __slots__ = ("_cache", "_coords", "_indexes", "_name", "_variable", "model")
+    labels: DataArray
+    model: Any = None
 
-    def __init__(self, *args, **kwargs):
-
-        # workaround until https://github.com/pydata/xarray/pull/5984 is merged
-        if isinstance(args[0], DataArray):
-            da = args[0]
-            args = (da.data, da.coords)
-            kwargs.update({"attrs": da.attrs, "name": da.name})
-
-        self.model = kwargs.pop("model", None)
-        super().__init__(*args, **kwargs)
-        assert self.name is not None, "Variable data does not have a name."
-
-    # We have to set the _reduce_method to None, in order to overwrite basic
-    # reduction functions as `sum`. There might be a better solution (?).
-    _reduce_method = None
 
     # Disable array function, only function defined below are supported
     # and set priority higher than pandas/xarray/numpy
     __array_ufunc__ = None
-    __array_priority__ = 10000
+    # __array_priority__ = 10000
 
     def __getitem__(self, keys) -> "ScalarVariable":
         keys = (keys,) if not isinstance(keys, tuple) else keys
@@ -121,18 +119,19 @@ class Variable(DataArray):
             "Set single values for each dimension in order to obtain a "
             "ScalarVariable. For all other purposes, use `.sel` and `.isel`."
         )
-        if not self.ndim:
+        if not self.labels.ndim:
             return ScalarVariable(self.data.item())
-        assert self.ndim == len(keys), f"expected {self.ndim} keys, got {len(keys)}."
-        key = dict(zip(self.dims, keys))
-        selector = [self.get_index(k).get_loc(v) for k, v in key.items()]
-        return ScalarVariable(self.data[tuple(selector)])
+        assert self.labels.ndim == len(keys), f"expected {self.labels.ndim} keys, got {len(keys)}."
+        key = dict(zip(self.labels.dims, keys))
+        selector = [self.labels.get_index(k).get_loc(v) for k, v in key.items()]
+        return ScalarVariable(self.labels.data[tuple(selector)])
 
+    @deprecated(details="Use `labels` instead of `to_array()`")
[…]
