### 2022-11-07 95fc5c4
Add monkey_patch work-around for left multiplication
