+++ linopy/common.py
@@ -6,7 +6,7 @@ Linopy common module.
 This module contains commonly used functions.
 """
 
-from functools import wraps
+from functools import wraps, update_wrapper, partial
 
 import numpy as np
 from xarray import DataArray, apply_ufunc, merge
@@ -113,5 +113,17 @@ def forward_as_properties(**routes):
             for attr in attrs:
                 add_accessor(cls, item, attr)
         return cls
+
+    return deco
+
+
+def monkey_patch(cls, pass_unpatched_method=False):
+    def deco(func):
+        wrapped = getattr(cls, func.__name__)
+        if pass_unpatched_method:
+            func = partial(func, unpatched_method=wrapped)
+        update_wrapper(func, wrapped)
+        setattr(cls, func.__name__, func)
+        return func
 
     return deco
