### 2022-11-07 cc58dd1
test_model_creation: Fix final test
