### 2022-11-07 a2607a4
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/keewis/blackdoc: v0.3.7 → v0.3.8](https://github.com/keewis/blackdoc/compare/v0.3.7...v0.3.8)
