+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
     -   id: docformatter
         args: [--in-place, --make-summary-multi-line, --pre-summary-newline]
 -   repo: https://github.com/keewis/blackdoc
-    rev: v0.3.7
+    rev: v0.3.8
     hooks:
       - id: blackdoc
 -   repo: https://github.com/MarcoGorelli/absolufy-imports
