### 2022-11-08 d164454
Move monkey patching into dedicated module
