+++ linopy/expressions.py
@@ -10,7 +10,7 @@ import functools
 import logging
 from dataclasses import dataclass
 from itertools import product, zip_longest
-from typing import Union
+from typing import Any, Hashable, Iterable, Mapping, Sequence, Union
 
 import numpy as np
 import pandas as pd
@@ -572,6 +572,8 @@ class LinearExpression:
     assign = exprwrap(Dataset.assign)
 
     assign_attrs = exprwrap(Dataset.assign_attrs)
+
+    assign_coords = exprwrap(Dataset.assign_coords)
 
     astype = exprwrap(Dataset.astype)
 
