### 2022-11-10 cfb4427
add LinearExpressionGroupby class
