### 2022-11-10 7f6ffa0
add LinearExpressionGroupby class
