### 2022-11-10 fd27342
add LinearExpressionRolling class adjust rename function
