### 2022-11-10 daa118f
add LinearExpressionRolling class adjust rename function
