+++ linopy/expressions.py
@@ -10,7 +10,7 @@ import functools
 import logging
 from dataclasses import dataclass
 from itertools import product, zip_longest
-from typing import Any, Hashable, Iterable, Mapping, Sequence, Union
+from typing import Any, Mapping, Union
 
 import numpy as np
 import pandas as pd
