### 2022-11-14 8e4130d
expression / constraints: allow returning None type when creating from rule
