### 2022-11-14 53f835d
expression / constraints: allow returning None type when creating from rule
