### 2022-11-16 #4823 @torressa
Hi, @FabianHofmann.

Sorry for the delay, I want to get the ball rolling on this one so this is my first pass on the review. I will come back with more once I find more time.

Really enjoyed looking through this, nicely written code and easy to follow.

I have mostly minor comments/things I'd like to see. Feel free to provide arguments why my comments are not needed/appropriate.

### Benchmarks
  - Comparison of model building times for some well-known model types: e.g. p-median, knapsack, n-queens. I understand that it may not be any different for `linopy`, however, other APIs may have a different (and significant overhead for different, more complex constraints).
  - I'd like to see a comparison against all major Python packages: PuLP, Pyomo, Highspy, Gurobipy (standard API), Gurobipy-matrixAPI (note major changes in v10.0 so please re-run benchmarks), ortools-pywraplp. (I may have missed some).
  - If you want to also include things like JuMP, (or other languages) it would be a more complete benchmark but I would say this is more optional.
  - Include solution retrieval in these benchmarks: i.e. extract all variable solution values.

### General code
 - Status codes, termination conditions, and constraint signs are not easily maintainable as a string (e.g. typos, `m.add_constraints(x, ",", 0)` would work just error when trying to solve. Please add some appropriate internal constants or some other data structure for these three things. Users can then export and use (similar to `GRB.OPTIMAL, GRB.EQUAL` etc).
 - What is the preferred way of building large models? `xarray` supports building from anything array-like, e.g. numpy, pandas dataframe, but would be good to know which is the best/fastest way of doing this, or your recommendation.

### General

- Please add the info you have in the docs pages (https://linopy.readthedocs.io/en/latest/) about contributing in a `CONTRIBUTING.md` file (this is standard GitHub).
- What's the status of generic non-linear constraints? Is this easy enough to do? https://github.com/PyPSA/linopy/issues/18
- PWL would be good (https://github.com/PyPSA/linopy/issues/19) but also other general constraints (not sure which ones would work). Also, not all solver support these.
- Functionality dive: for someone not used to the xarray syntax, it would be cool that you showcase some simple examples already in the README, however, this is my preference. I would also like to see some more complex examples in the examples directory. I like the shifting stuff with the `y` variables for example.

### File Types
- More as a question: I have never heard of NetCDF, what are the advantages when compared to the widely used `mps` format, for example?
- I would like to see functionality for MPS as well. And tested for across all solvers.
- LP files contain unnecessary blank lines. These can make files a lot bigger in size than they need to be.
- Some files are executable: `setup.py`, `linopy/__init__.py`, `test/test_model_creation.py`.

### Paper edits
 - PuLP needs a proper citation.
 - N-dimensional labelled is not very well known to the OR community (I'd never heard this term). I would revise the title to something that is more in the court of the core-users of this package. (again, this is my opinion).
