### 2022-11-30 0e9e04b
Define field as default_factory to specify mutable default value
