### 2022-11-30 2ad8d56
Define field as default_factory to specify mutable default value
