### 2022-12-02 9f7c58d
fix minor typos in doc strings
