### 2022-12-02 c88bf62
fix minor typos in doc strings
