### 2022-12-04 ca491d0
setup.py: use better version spec for readthedocs
