### 2022-12-04 bb381c2
setup.py follow up
