### 2022-12-04 38d316c
setup.py: follow up
