### 2022-12-07 0b3eeb4
setup.py: include highspy
