+++ setup.py
@@ -55,8 +55,7 @@ setup(
             "pre-commit",
             "paramiko",
             "gurobipy",
-            # until available for windows/mac
-            # "highspy",
+            "highspy",
             "cplex",
             "xpress",
         ],
