### 2022-12-09 beffc90
make MPS export dependent on whether highspy is available
