### 2022-12-09 66404bd
test_io.py: test mps export only when highspy is installed
