### 2022-12-09 3222b60
test: modularize test_io
