### 2022-12-09 1118a03
test: modularize test_variable
