+++ test/test_variable.py
@@ -10,15 +10,41 @@ import numpy as np
 import pandas as pd
 import pytest
 import xarray as xr
+from xarray.testing import assert_equal
 
 import linopy
 from linopy import Model
 
 
-def test_variable_getter():
+@pytest.fixture
+def m():
     m = Model()
-    x = m.add_variables(coords=[range(10)])
+    x = m.add_variables(coords=[pd.RangeIndex(10, name="first")], name="x")
+    y = m.add_variables(coords=[pd.Index([1, 2, 3], name="second")], name="y")
+    z = m.add_variables(0, 10, name="z")
+    return m
 
+
+@pytest.fixture
+def x(m):
+    return m.variables["x"]
+
+
+@pytest.fixture
+def z(m):
+    return m.variables["z"]
+
+
+def test_variables_nvars(m):
+    assert m.variables.nvars == 14
+
+    idx = pd.RangeIndex(10, name="first")
+    mask = pd.Series([True] * 5 + [False] * 5, idx)
+    m.add_variables(coords=[idx], mask=mask)
+    assert m.variables.nvars == 19
+
+
+def test_variable_getter(x):
     assert isinstance(x[0], linopy.variables.ScalarVariable)
 
     with pytest.raises(AssertionError):
@@ -30,111 +56,38 @@ def test_variable_getter():
     with pytest.raises(AssertionError):
         x[[1, 2, 3]]
 
+
+def test_variable_loc(x):
     assert isinstance(x.loc[[1, 2, 3]], linopy.Variable)
 
 
-def test_variable_repr():
-    m = Model()
-    m.variables.__repr__()
+def test_variable_sel(x):
+    assert isinstance(x.sel(first=[1, 2, 3]), linopy.Variable)
 
-    x = m.add_variables()
+
+def test_variable_isel(x):
+    assert isinstance(x.isel(first=[1, 2, 3]), linopy.Variable)
+    assert_equal(
+        x.isel(first=[0, 1]).labels,
+        x.sel(first=[0, 1]).labels,
+    )
+
+
+def test_variable_repr(x):
     x.__repr__()
     x._repr_html_()
 
-    m.variables.__repr__()
 
-    y = m.add_variables(coords=[pd.Index([1, 2, 3], name="time")], name="y")
-    y.__repr__()
-    y._repr_html_()
-
+def test_variables_repr(m):
     m.variables.__repr__()
 
 
-def test_variable_bound_accessor():
-    m = Model()
-    x = m.add_variables(0, 10)
-    assert x.upper.item() == 10
-    assert x.lower.item() == 0
+def test_variable_upper_getter(z):
+    assert z.upper.item() == 10
 
 
-def test_variable_modification():
-    m = Model()
-    x = m.add_variables(0, 10)
-    x.upper = 20
-    assert x.upper.item() == 20
-
-    x.lower = 8
-    assert x.lower == 8
-
-
-def test_variable_modification_M():
-    m = Model()
-    lower = pd.Series(0, index=range(10))
-    upper = pd.Series(range(10, 20), index=range(10))
-    x = m.add_variables(lower, upper)
-
-    new_upper = pd.Series(range(25, 35), index=range(10))
-    x.upper = new_upper
-    assert isinstance(x.upper, xr.DataArray)
-    assert (x.upper == new_upper).all()
-
-    new_lower = pd.Series(range(15, 25), index=range(10))
-    x.lower = new_lower
-    assert isinstance(x.lower, xr.DataArray)
-    assert (x.lower == new_lower).all()
-
-
-def test_variable_sum():
-    m = Model()
-    x = m.add_variables(coords=[range(10)])
-    res = x.sum()
-    assert res.nterm == 10
-
-
-def test_nvars():
-    m = Model()
-    m.add_variables(coords=[range(10)])
[…]
