### 2022-12-09 d7f89ca
test: modularize test_variable
