### 2022-12-09 da30081
test_contraint: modularize/split functions
