+++ test/test_constraint.py
@@ -17,106 +17,137 @@ from linopy import Model
 from linopy.constraints import AnonymousConstraint
 
 
-def test_constraint_repr():
+@pytest.fixture
+def m():
     m = Model()
+    x = m.add_variables(coords=[pd.RangeIndex(10, name="first")], name="x")
+    y = m.add_variables(coords=[pd.Index([1, 2, 3], name="second")], name="y")
+    z = m.add_variables(0, 10, name="z")
+    c = m.add_constraints(x >= 0, name="c")
+    return m
 
-    x = m.add_variables()
-    c = m.add_constraints(x, ">=", 0)
+
+@pytest.fixture
+def x(m):
+    return m.variables["x"]
+
+
+@pytest.fixture
+def y(m):
+    return m.variables["y"]
+
+
+@pytest.fixture
+def c(m):
+    return m.constraints["c"]
+
+
+def test_constraint_repr(c):
     c.__repr__()
     c._repr_html_()
 
 
-def test_constraints_repr():
-    m = Model()
-    m.constraints.__repr__()
-    x = m.add_variables()
-    m.add_constraints(x, ">=", 0)
+def test_constraints_repr(m):
     m.constraints.__repr__()
 
 
-def test_scalarconstraint():
-    m = Model()
-    coords = [pd.Index(range(10))]
-    x = m.add_variables(coords=coords)
-    con = m.add_constraints(x[0] >= 0)
-    assert isinstance(con, linopy.constraints.Constraint)
-
-    con = m.add_constraints(x[0] + x[1] >= 0)
-    assert isinstance(con, linopy.constraints.Constraint)
-
-    con = m.add_constraints(x[0], ">=", 0)
-    assert isinstance(con, linopy.constraints.Constraint)
-
-    con = m.add_constraints(x[0] + x[1], ">=", 0)
-    assert isinstance(con, linopy.constraints.Constraint)
+def test_empty_constraints_repr():
+    # test empty contraints
+    Model().constraints.__repr__()
 
 
-def test_from_rule():
-    m = Model()
-    coords = pd.RangeIndex(10), ["a", "b"]
-    x = m.add_variables(0, 100, coords)
+def test_constraints_getter(m, c):
+    assert c.shape == (10,)
+    assert isinstance(m.constraints[["c"]], linopy.constraints.Constraints)
 
+
+def test_constraint_from_rule(m, x, y):
     def bound(m, i, j):
         if i % 2:
-            return (i - 1) * x[i - 1, j] >= 0
+            return (i - 1) * x[i - 1] + y[j] >= 0
         else:
-            return i * x[i, j] >= 0
+            return i * x[i] >= 0
 
+    coords = [x.coords["first"], y.coords["second"]]
     con = AnonymousConstraint.from_rule(m, bound, coords)
     assert isinstance(con, AnonymousConstraint)
-    assert con.lhs.nterm == 1
+    assert con.lhs.nterm == 2
 
+
+def test_constraint_from_rule_with_none_return(m, x, y):
     def bound(m, i, j):
         if i % 2:
-            return i * x[i, j] >= 0
+            return i * x[i] + y[j] >= 0
 
+    coords = [x.coords["first"], y.coords["second"]]
     con = AnonymousConstraint.from_rule(m, bound, coords)
     assert isinstance(con, AnonymousConstraint)
-    assert con.lhs.nterm == 1
+    assert con.lhs.nterm == 2
     assert (con.lhs.vars.loc[0, :] == -1).all()
     assert (con.lhs.vars.loc[1, :] != -1).all()
 
 
-def test_constraint_accessor():
-    m = Model()
-    x = m.add_variables()
-    y = m.add_variables()
[…]
