### 2022-12-09 83419ee
test linear expression: refactor code; apply flake8 hints to some other tests
