+++ test/test_linear_expression.py
@@ -16,25 +16,54 @@ from linopy import LinearExpression, Model, merge
 from linopy.expressions import ScalarLinearExpression
 from linopy.testing import assert_linequal
 
-m = Model()
 
-x = m.add_variables(pd.Series([0, 0]), 1, name="x")
-y = m.add_variables(4, pd.Series([8, 10]), name="y")
-z = m.add_variables(0, pd.DataFrame([[1, 2], [3, 4], [5, 6]]).T, name="z")
-v = m.add_variables(coords=[pd.RangeIndex(20, name="dim_2")], name="v")
+@pytest.fixture
+def m():
+    m = Model()
 
-idx = pd.MultiIndex.from_product([[1, 2], ["a", "b"]], names=("level1", "level2"))
-idx.name = "dim_3"
-u = m.add_variables(coords=[idx], name="u")
+    m.add_variables(pd.Series([0, 0]), 1, name="x")
+    m.add_variables(4, pd.Series([8, 10]), name="y")
+    m.add_variables(0, pd.DataFrame([[1, 2], [3, 4], [5, 6]]).T, name="z")
+    m.add_variables(coords=[pd.RangeIndex(20, name="dim_2")], name="v")
+
+    idx = pd.MultiIndex.from_product([[1, 2], ["a", "b"]], names=("level1", "level2"))
+    idx.name = "dim_3"
+    m.add_variables(coords=[idx], name="u")
+    return m
 
 
-def test_repr():
+@pytest.fixture
+def x(m):
+    return m.variables["x"]
+
+
+@pytest.fixture
+def y(m):
+    return m.variables["y"]
+
+
+@pytest.fixture
+def z(m):
+    return m.variables["z"]
+
+
+@pytest.fixture
+def v(m):
+    return m.variables["v"]
+
+
+@pytest.fixture
+def u(m):
+    return m.variables["u"]
+
+
+def test_repr(m):
     expr = m.linexpr((10, "x"), (1, "y"))
     expr.__repr__()
     expr._repr_html_()
 
 
-def test_values():
+def test_linexpr_with_scalars(m):
     expr = m.linexpr((10, "x"), (1, "y"))
     target = xr.DataArray(
         [[10, 1], [10, 1]], coords={"dim_0": [0, 1]}, dims=["dim_0", "_term"]
@@ -42,12 +71,24 @@ def test_values():
     assert_equal(expr.coeffs, target)
 
 
-def test_duplicated_index():
+def test_linexpr_with_series(m, v):
+    lhs = pd.Series(np.arange(20)), v
+    expr = m.linexpr(lhs)
+    isinstance(expr, LinearExpression)
+
+
+def test_linexpr_with_dataframe(m, v):
+    lhs = pd.DataFrame({"a": np.arange(20)}).T, v
+    expr = m.linexpr(lhs)
+    isinstance(expr, LinearExpression)
+
+
+def test_linexpr_duplicated_index(m):
     expr = m.linexpr((10, "x"), (-1, "x"))
     assert (expr.data._term == [0, 1]).all()
 
 
-def test_variable_to_linexpr():
+def test_linear_expression_with_multiplication(x):
     expr = 1 * x
     assert isinstance(expr, LinearExpression)
     assert expr.nterm == 1
@@ -62,6 +103,11 @@ def test_variable_to_linexpr():
     expr = x / 1.0
     assert isinstance(expr, LinearExpression)
 
+    expr = np.array([1, 2]) * x
+    assert isinstance(expr, LinearExpression)
+
+
+def test_linear_expression_with_addition(m, x, y):
     expr = 10 * x + y
     assert isinstance(expr, LinearExpression)
     assert_linequal(expr, m.linexpr((10, "x"), (1, "y")))
@@ -74,6 +120,8 @@ def test_variable_to_linexpr():
[…]
