### 2022-12-09 13eb0d2
test linear expression: refactor code; apply flake8 hints to some other tests
