### 2022-12-10 6210a17
run flake8 on all test modularize model creation
