### 2022-12-10 4dd036a
run flake8 on all test modularize model creation
