### 2022-12-10 ef90ddf
support integer variables
