### 2022-12-11 84afb10
add constants module support alternative signs add checks for signs
