+++ test/test_constraints.py
@@ -12,7 +12,7 @@ import pandas as pd
 import pytest
 import xarray as xr
 
-from linopy import Model
+from linopy import EQUAL, GREATER_EQUAL, LESS_EQUAL, Model
 
 # Test model functions
 
@@ -25,7 +25,7 @@ def test_constraint_assignment():
     x = m.add_variables(lower, upper)
     y = m.add_variables()
 
-    m.add_constraints(1 * x + 10 * y, "=", 0)
+    m.add_constraints(1 * x + 10 * y, EQUAL, 0)
 
     for attr in m.constraints.dataset_attrs:
         assert "con0" in getattr(m.constraints, attr)
@@ -65,7 +65,7 @@ def test_constraint_assignment_with_tuples():
     x = m.add_variables(lower, upper)
     y = m.add_variables()
 
-    m.add_constraints([(1, x), (10, y)], "=", 0, name="c")
+    m.add_constraints([(1, x), (10, y)], EQUAL, 0, name="c")
     for attr in m.constraints.dataset_attrs:
         assert "c" in getattr(m.constraints, attr)
     assert m.constraints.labels.c.shape == (10, 10)
@@ -77,7 +77,7 @@ def test_constraint_assignment_chunked():
     lower = pd.DataFrame(np.zeros((10, 10)))
     upper = pd.Series(np.ones((10)))
     x = m.add_variables(lower, upper)
-    m.add_constraints(x, ">=", 0, name="c")
+    m.add_constraints(x, GREATER_EQUAL, 0, name="c")
     assert m.constraints.coeffs.c.data.shape == (
         10,
         10,
@@ -86,20 +86,13 @@ def test_constraint_assignment_chunked():
     assert isinstance(m.constraints.coeffs.c.data, dask.array.core.Array)
 
 
-def test_wrong_constraint_assignment_wrong_sign():
-    m = Model()
-    x = m.add_variables()
-    with pytest.raises(ValueError):
-        m.add_constraints(x, "==", 0)
-
-
 def test_wrong_constraint_assignment_repeated():
     # repeated variable assignment is forbidden
     m = Model()
     x = m.add_variables()
-    m.add_constraints(x, "<=", 0, name="con")
+    m.add_constraints(x, LESS_EQUAL, 0, name="con")
     with pytest.raises(ValueError):
-        m.add_constraints(x, "<=", 0, name="con")
+        m.add_constraints(x, LESS_EQUAL, 0, name="con")
 
 
 def test_masked_constraints():
@@ -111,6 +104,6 @@ def test_masked_constraints():
     y = m.add_variables()
 
     mask = pd.Series([True] * 5 + [False] * 5)
-    m.add_constraints(1 * x + 10 * y, "=", 0, mask=mask)
+    m.add_constraints(1 * x + 10 * y, EQUAL, 0, mask=mask)
     assert (m.constraints.labels.con0[0:5, :] != -1).all()
     assert (m.constraints.labels.con0[5:10, :] == -1).all()
