+++ test/test_inconsistency_checks.py
@@ -9,7 +9,7 @@ Created on Mon Feb 28 15:47:32 2022.
 import numpy as np
 import pytest
 
-from linopy import Model
+from linopy import GREATER_EQUAL, Model
 
 
 def test_nan_in_variable_lower():
@@ -18,8 +18,8 @@ def test_nan_in_variable_lower():
     x = m.add_variables(lower=np.nan, name="x")
     y = m.add_variables(name="y")
 
-    m.add_constraints(2 * x + 6 * y, ">=", 10)
-    m.add_constraints(4 * x + 2 * y, ">=", 3)
+    m.add_constraints(2 * x + 6 * y, GREATER_EQUAL, 10)
+    m.add_constraints(4 * x + 2 * y, GREATER_EQUAL, 3)
 
     m.add_objective(2 * y + x)
     with pytest.raises(ValueError):
@@ -32,8 +32,8 @@ def test_nan_in_variable_upper():
     x = m.add_variables(upper=np.nan, name="x")
     y = m.add_variables(name="y")
 
-    m.add_constraints(2 * x + 6 * y, ">=", 10)
-    m.add_constraints(4 * x + 2 * y, ">=", 3)
+    m.add_constraints(2 * x + 6 * y, GREATER_EQUAL, 10)
+    m.add_constraints(4 * x + 2 * y, GREATER_EQUAL, 3)
 
     m.add_objective(2 * y + x)
     with pytest.raises(ValueError):
@@ -46,8 +46,8 @@ def test_nan_in_constraint_coeffs():
     x = m.add_variables(name="x")
     y = m.add_variables(name="y")
 
-    m.add_constraints(np.nan * x + 6 * y, ">=", 10)
-    m.add_constraints(4 * x + 2 * y, ">=", 3)
+    m.add_constraints(np.nan * x + 6 * y, GREATER_EQUAL, 10)
+    m.add_constraints(4 * x + 2 * y, GREATER_EQUAL, 3)
 
     m.add_objective(2 * y + x)
     with pytest.raises(ValueError):
@@ -60,12 +60,8 @@ def test_nan_in_constraint_sign():
     x = m.add_variables(name="x")
     y = m.add_variables(name="y")
 
-    m.add_constraints(2 * x + 6 * y, np.nan, 10)
-    m.add_constraints(4 * x + 2 * y, ">=", 3)
-
-    m.add_objective(2 * y + x)
     with pytest.raises(ValueError):
-        m.solve()
+        m.add_constraints(2 * x + 6 * y, np.nan, 10)
 
 
 def test_nan_in_constraint_rhs():
@@ -74,8 +70,8 @@ def test_nan_in_constraint_rhs():
     x = m.add_variables(name="x")
     y = m.add_variables(name="y")
 
-    m.add_constraints(2 * x + 6 * y, ">=", np.nan)
-    m.add_constraints(4 * x + 2 * y, ">=", 3)
+    m.add_constraints(2 * x + 6 * y, GREATER_EQUAL, np.nan)
+    m.add_constraints(4 * x + 2 * y, GREATER_EQUAL, 3)
 
     m.add_objective(2 * y + x)
     with pytest.raises(ValueError):
@@ -88,8 +84,8 @@ def test_nan_in_objective():
     x = m.add_variables(name="x")
     y = m.add_variables(name="y")
 
-    m.add_constraints(2 * x + 6 * y, ">=", np.nan)
-    m.add_constraints(4 * x + 2 * y, ">=", 3)
+    m.add_constraints(2 * x + 6 * y, GREATER_EQUAL, np.nan)
+    m.add_constraints(4 * x + 2 * y, GREATER_EQUAL, 3)
 
     m.add_objective(np.nan * y + x)
     with pytest.raises(ValueError):
