### 2022-12-12 93d6a5c
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 22.10.0 → 22.12.0](https://github.com/psf/black/compare/22.10.0...22.12.0)
- [github.com/PyCQA/isort: 5.10.1 → 5.11.0](https://github.com/PyCQA/isort/compare/5.10.1...5.11.0)
