+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
     -   id: trailing-whitespace
     -   id: check-merge-conflict
 -   repo: https://github.com/psf/black
-    rev: 22.10.0
+    rev: 22.12.0
     hooks:
     -   id: black
 -   repo: https://github.com/PyCQA/docformatter
@@ -23,7 +23,7 @@ repos:
     hooks:
     -   id: absolufy-imports
 -   repo: https://github.com/PyCQA/isort
-    rev: 5.10.1
+    rev: 5.11.0
     hooks:
       - id: isort
         args: ["--profile", "black", "--filter-files"]
