### 2022-12-12 90f0fd0
define fix set of status and termination conditions
