+++ linopy/constants.py
@@ -3,6 +3,16 @@
 """
 Linopy module for defining constant values used within the package.
 """
+import logging
+from dataclasses import dataclass, field
+from enum import Enum
+from typing import Any, Optional, Tuple, Union
+
+import numpy as np
+import pandas as pd
+
+logger = logging.getLogger(__name__)
+
 
 EQUAL = "="
 GREATER_EQUAL = ">="
@@ -21,3 +31,152 @@ sign_replace_dict = {
     short_GREATER_EQUAL: GREATER_EQUAL,
     short_LESS_EQUAL: LESS_EQUAL,
 }
+
+
+class SolverStatus(Enum):
+    ok = "ok"
+    warning = "warning"
+    error = "error"
+    aborted = "aborted"
+    unknown = "unknown"
+
+    @classmethod
+    def process(cls, status: str) -> "SolverStatus":
+        try:
+            return cls(status)
+        except ValueError:
+            return cls("unknown")
+
+    @classmethod
+    def from_termination_condition(
+        cls, termination_condition: "TerminationCondition"
+    ) -> "SolverStatus":
+        for status in STATUS_TO_TERMINATION_CONDITION_MAP:
+            if termination_condition in STATUS_TO_TERMINATION_CONDITION_MAP[status]:
+                return status
+        return cls.unknown
+
+
+class TerminationCondition(Enum):
+    # UNKNOWN
+    unknown = "unknown"
+
+    # OK
+    optimal = "optimal"
+    time_limit = "time_limit"
+    iteration_limit = "iteration_limit"
+    terminated_by_limit = "terminated_by_limit"
+    suboptimal = "suboptimal"
+
+    # WARNING
+    unbounded = "unbounded"
+    infeasible = "infeasible"
+    infeasible_or_unbounded = "infeasible_or_unbounded"
+    other = "other"
+
+    # ERROR
+    internal_solver_error = "internal_solver_error"
+    error = "error"
+
+    # ABORTED
+    user_interrupt = "user_interrupt"
+    resource_interrupt = "resource_interrupt"
+    licensing_problems = "licensing_problems"
+
+    @classmethod
+    def process(cls, termination_condition: str) -> "TerminationCondition":
+        try:
+            return cls(termination_condition)
+        except ValueError:
+            return cls("unknown")
+
+
+STATUS_TO_TERMINATION_CONDITION_MAP = {
+    SolverStatus.ok: [
+        TerminationCondition.optimal,
+        TerminationCondition.iteration_limit,
+        TerminationCondition.time_limit,
+        TerminationCondition.terminated_by_limit,
+        TerminationCondition.other,
+        TerminationCondition.suboptimal,
+    ],
+    SolverStatus.warning: [
+        TerminationCondition.unbounded,
+        TerminationCondition.infeasible,
+        TerminationCondition.infeasible_or_unbounded,
+        TerminationCondition.other,
+    ],
+    SolverStatus.error: [
+        TerminationCondition.internal_solver_error,
+        TerminationCondition.error,
+    ],
+    SolverStatus.aborted: [
+        TerminationCondition.user_interrupt,
+        TerminationCondition.resource_interrupt,
+        TerminationCondition.licensing_problems,
+    ],
+    SolverStatus.unknown: [TerminationCondition.unknown],
+}
+
+
+@dataclass
+class Status:
+    status: SolverStatus
+    termination_condition: TerminationCondition
+    legacy_status: str = ""
+
+    @classmethod
+    def process(cls, status: str, termination_condition: str) -> "Status":
+        return cls(
+            status=SolverStatus.process(status),
+            termination_condition=TerminationCondition.process(termination_condition),
+            legacy_status=(status, termination_condition),
+        )
+
+    @classmethod
+    def from_termination_condition(
[…]
