### 2022-12-14 41e9827
rename get_solution to safe_get_solution
