### 2022-12-14 55d53f0
add Model.solve to api ref
