### 2022-12-16 bf67ce5
add SCIP solver interface
