### 2022-12-16 0fc98c4
update solver installation guideline
