### 2022-12-19 ead6259
filter out very small coefficients
