### 2022-12-19 5e198e9
improve representation of infeasible constraints
