### 2022-12-21 9aa13d0
rewrite repr for variables
