### 2022-12-24 4afb37d
add chi a label indicator; rewrite expressions repr
