### 2022-12-29 3f3576e
refactor display of constraints and anonymous constraints
