+++ test/test_repr.py
@@ -13,6 +13,16 @@ x = m.add_variables(lower, 10, coords=[lower.index], name="x")
 y = m.add_variables(0, 10, name="y")
 z = m.add_variables(name="z", binary=True)
 a = m.add_variables(coords=[lower.index], name="a", binary=True)
+b = m.add_variables(coords=[lower.index], name="b", integer=True)
+
+# create constraint for each variable
+cu = m.add_constraints(1 * u >= 0, name="cu")
+cv = m.add_constraints(1 * v >= 0, name="cv")
+cx = m.add_constraints(1 * x >= 0, name="cx")
+cy = m.add_constraints(1 * y >= 0, name="cy")
+cz = m.add_constraints(1 * z >= 0, name="cz")
+ca = m.add_constraints(1 * a >= 0, name="ca")
+cav = m.add_constraints(1 * a + 1 * v >= 0, name="cav")
 
 
 def test_variable_repr_u():
@@ -55,6 +65,10 @@ def test_linear_expression_y():
     repr(y.to_linexpr())
 
 
+def test_linear_expression_long():
+    repr(x.sum())
+
+
 def test_constraint_u():
     repr(u >= 0)
 
@@ -69,3 +83,27 @@ def test_constraint_x():
 
 def test_constraint_y():
     repr(y >= 0)
+
+
+def test_constraint_cu():
+    repr(cu)
+
+
+def test_constraint_cv():
+    repr(cv)
+
+
+def test_constraint_cx():
+    repr(cx)
+
+
+def test_constraint_cy():
+    repr(cy)
+
+
+def test_constraint_cz():
+    repr(cz)
+
+
+def test_constraint_ca():
+    repr(ca)
