+++ linopy/constraints.py
@@ -22,15 +22,22 @@ from xarray import DataArray, Dataset
 from linopy import expressions, variables
 from linopy.common import (
     _merge_inplace,
+    forward_as_properties,
     has_optimized_model,
+    head_tail_range,
     is_constant,
     maybe_replace_signs,
+    print_single_constraint_label,
+    print_single_expression,
     replace_by_map,
 )
 from linopy.constants import EQUAL, GREATER_EQUAL, LESS_EQUAL
 
 
 @dataclass(repr=False)
+@forward_as_properties(
+    lhs=["attrs", "coords", "dims", "indexes", "nterm", "ndim", "size"]
+)
 class Constraint:
     """
     Projection to a single constraint in a model.
@@ -39,29 +46,60 @@ class Constraint:
     functions can be applied to it.
     """
 
-    name: str
-    model: Any
+    labels: DataArray = field(default_factory=DataArray)
+    model: Any = None
+    name: str = None
 
     def __repr__(self):
         """
-        Get the string representation of the constraints.
+        Get the string representation of the Constraint.
         """
-        data_string = "Constraint labels:\n" + self.labels.__repr__().split("\n", 1)[1]
-        extend_line = "-" * len(self.name)
-        return (
-            f"Constraint '{self.name}':\n"
-            f"--------------{extend_line}\n\n"
-            f"{data_string}"
-        )
 
-    def _repr_html_(self):
-        """
-        Get the html representation of the variables.
-        """
-        # return self.__repr__()
-        data_string = self.labels._repr_html_()
-        data_string = data_string.replace("xarray.DataArray", "linopy.Constraint")
-        return data_string
+        # return single if only one exist
+        if self.size == self.nterm:
+            expr_string = print_single_expression(
+                self.lhs.coeffs.values, self.lhs.vars.values
+            )
+            label_string = print_single_constraint_label(self.labels)
+            header = "Constraint:\n-----------"
+            return f"{header}\n{expr_string} {self.sign.item()} {self.rhs.item()} \t| {label_string}"
+
+        # print only a few values
+        nexprs = self.size // self.nterm
+        max_prints = 14
+        split_at = max_prints // 2
+        to_print = head_tail_range(nexprs, max_prints)
+        coords = self.lhs.unravel_coords(to_print)
+
+        # loop over all values to print
+        data_string = ""
+        for i, coord in enumerate(coords):
+            # sign and rhs might only be defined for some dimensions
+            sign_coord = tuple(
+                c for i, c in enumerate(coord) if list(self.dims)[i] in self.sign.dims
+            )
+            rhs_coord = tuple(
+                c for i, c in enumerate(coord) if list(self.dims)[i] in self.rhs.dims
+            )
+
+            coord_string = "[" + ", ".join([str(c) for c in coord]) + "]"
+            expr_string = print_single_expression(
+                self.coeffs[coord].values, self.vars[coord].values
+            )
+            sign_string = f"{self.sign[sign_coord].values}"
+            rhs_string = f"{self.rhs[rhs_coord].values}"
+            label_string = print_single_constraint_label(self.labels[coord].values)
+
+            data_string += f"\n{coord_string}:  {expr_string} {sign_string} {rhs_string} | {label_string}"
+
+            if i == split_at - 1 and nexprs > max_prints:
+                data_string += "\n\t\t..."
+
+        # create shape string
+        nonterm_dims = [(k, v) for k, v in self.dims.items() if not k.startswith("_")]
+        shape_string = "(" + ", ".join([f"{k}: {v}" for k, v in nonterm_dims]) + ")"
[…]

[…]
