### 2023-01-09 cc2d230
display vars with coordinates in __repr__
