+++ linopy/expressions.py
@@ -38,7 +38,7 @@ def exprwrap(method, *default_args, **new_default_kwargs):
         for k, v in new_default_kwargs.items():
             kwargs.setdefault(k, v)
         return expr.__class__(
-            expr.model, method(_expr_unwrap(expr), *default_args, *args, **kwargs)
+            method(_expr_unwrap(expr), *default_args, *args, **kwargs), expr.model
         )
 
     _exprwrap.__doc__ = f"Wrapper for the xarray {method} function for linopy.Variable"
@@ -70,7 +70,7 @@ class LinearExpressionGroupby:
 
     def map(self, func, shortcut=False, args=(), **kwargs):
         return LinearExpression(
-            self.model, self.groupby.map(func, shortcut=shortcut, args=args, **kwargs)
+            self.groupby.map(func, shortcut=shortcut, args=args, **kwargs), self.model
         )
 
     def sum(self, **kwargs):
@@ -102,7 +102,7 @@ class LinearExpressionRolling:
             .stack(_term=["_stacked_term", "_rolling_term"])
             .reset_index("_term", drop=True)
         )
-        return LinearExpression(self.model, ds)
+        return LinearExpression(ds, self.model)
 
 
 @forward_as_properties(data=["attrs", "coords", "indexes"])
@@ -144,29 +144,42 @@ class LinearExpression:
     <class 'linopy.expressions.LinearExpression'>
     """
 
-    __slots__ = ("data", "model")
-
-    fill_value = {"vars": -1, "coeffs": np.nan}
-
+    __slots__ = ("_data", "_model")
     __array_ufunc__ = None
     __array_priority__ = 10000
 
-    def __init__(self, model, data_vars=None, coords=None, attrs=None):
-        ds = Dataset(data_vars, coords, attrs)
+    _fill_value = {"vars": -1, "coeffs": np.nan}
 
-        if not len(ds):
-            vars = DataArray(np.array([], dtype=int), dims="_term")
-            coeffs = DataArray(np.array([], dtype=float), dims="_term")
-            ds = ds.assign(coeffs=coeffs, vars=vars)
+    def __init__(self, data, model):
+        from linopy.model import Model
 
-        assert set(ds).issuperset({"coeffs", "vars"})
-        if np.issubdtype(ds.vars, np.floating):
-            ds["vars"] = ds.vars.fillna(-1).astype(int)
-        (ds,) = xr.broadcast(ds)
-        ds = ds.transpose(..., "_term")
+        if data is None:
+            da = xr.DataArray([], dims=["_term"])
+            data = Dataset({"coeffs": da, "vars": da})
 
-        self.model = model
-        self.data = ds
+        if not isinstance(data, Dataset):
+            raise ValueError(
+                f"data must be an instance of xarray.Dataset, got {type(data)}"
+            )
+
+        if not set(data).issuperset({"coeffs", "vars"}):
+            raise ValueError("data must contain the variables 'coeffs' and 'vars'")
+
+        # make sure that all non-helper dims have coordinates
+        if not all(dim in data.coords for dim in data.dims if not dim.startswith("_")):
+            raise ValueError("data must have coordinates for all non-helper dimensions")
+
+        if np.issubdtype(data.vars, np.floating):
+            data["vars"] = data.vars.fillna(-1).astype(int)
+
+        (data,) = xr.broadcast(data)
+        data = data.transpose(..., "_term")
+
+        if not isinstance(model, Model):
+            raise ValueError("model must be an instance of linopy.Model")
+
+        self._model = model
+        self._data = data
 
     def unravel_coords(self, index):
         idx = np.unravel_index(index, self.shape[:-1])
[…]
