+++ linopy/io.py
@@ -519,7 +519,7 @@ def read_netcdf(path, **kwargs):
 
     for attr in m.dataset_attrs:
         setattr(m, attr, get_and_rename(all_ds, attr))
-    m._objective = LinearExpression(m, get_and_rename(all_ds, "objective"))
+    m._objective = LinearExpression(get_and_rename(all_ds, "objective"), m)
 
     for k in m.scalar_attrs:
         setattr(m, k, all_ds.attrs.pop(k))
