### 2023-01-10 2118cf7
apply new repr to Scalar classes
