+++ linopy/model.py
@@ -25,7 +25,6 @@ from linopy.constraints import (
     AnonymousScalarConstraint,
     Constraints,
 )
-from linopy.eval import Expr
 from linopy.expressions import LinearExpression, ScalarLinearExpression
 from linopy.io import to_block_files, to_file, to_gurobipy, to_highspy, to_netcdf
 from linopy.matrices import MatrixAccessor
@@ -407,18 +406,18 @@ class Model:
         >>> m = Model()
         >>> time = pd.RangeIndex(10, name="Time")
         >>> m.add_variables(lower=0, coords=[time], name="x")
-        Variable 'x':
-        -------------
-        +0 ≤  x[0] ≤ inf
-        +0 ≤  x[1] ≤ inf
-        +0 ≤  x[2] ≤ inf
-        +0 ≤  x[3] ≤ inf
-        +0 ≤  x[4] ≤ inf
-        +0 ≤  x[5] ≤ inf
-        +0 ≤  x[6] ≤ inf
-        +0 ≤  x[7] ≤ inf
-        +0 ≤  x[8] ≤ inf
-        +0 ≤  x[9] ≤ inf
+        Continuous Variable (Time: 10)
+        ------------------------------
+        0 ≤  x[0] ≤ inf
+        0 ≤  x[1] ≤ inf
+        0 ≤  x[2] ≤ inf
+        0 ≤  x[3] ≤ inf
+        0 ≤  x[4] ≤ inf
+        0 ≤  x[5] ≤ inf
+        0 ≤  x[6] ≤ inf
+        0 ≤  x[7] ≤ inf
+        0 ≤  x[8] ≤ inf
+        0 ≤  x[9] ≤ inf
         """
         if name is None:
             name = "var" + str(self._varnameCounter)
@@ -828,176 +827,21 @@ class Model:
         else:
             raise TypeError(f"Not supported type {args}.")
 
-    def _eval(self, expr: str, **kwargs):
-        from pandas.core.computation.eval import eval as pd_eval
-
-        kwargs.setdefault("engine", "python")
-        resolvers = kwargs.pop("resolvers", None)
-        kwargs["level"] = kwargs.pop("level", 0) + 1
-        resolvers = [self.variables, self.parameters]
-        kwargs["resolvers"] = kwargs.get("resolvers", ()) + tuple(resolvers)
-        return pd_eval(expr, inplace=False, **kwargs)
-
     def vareval(self, expr: str, eval_kw=None, **kwargs):
-        """
-        Define a variable based a string expression (experimental).
-
-        The function mirrors the behavior of `pandas.DataFrame.eval()`, e.g.
-        global variables can be referenced with a @-suffix, model attributes
-        such as parameters and variables can be referenced by the key.
-
-        Parameters
-        ----------
-        expr : str
-            Valid string to be compiled as a variable definition
-            (lower and upper bounds!).
-        eval_kw : dict
-            Keyword arguments to be passed to `pandas.eval`.
-        **kwargs :
-            Keyword arguments to be passed to `model.add_constraints`.
-
-
-        Returns
-        -------
-        linopy.Variable
-            Variable which was added to the model.
-
-
-        Examples
-        --------
-
-        >>> import linopy, xarray as xr
-        >>> m = linopy.Model()
-        >>> lower = xr.DataArray(np.zeros((10, 10)), coords=[range(10), range(10)])
-        >>> upper = xr.DataArray(np.ones((10, 10)), coords=[range(10), range(10)])
-        >>> x = m.vareval("@lower <= x <= @upper")  # doctest:+SKIP
-
-        This is the same as
-
-        >>> x = m.add_variables(lower, upper, name="x")
-        """
[…]
