### 2023-01-10 c562220
adjust header of linear expression repr
