+++ linopy/solvers.py
@@ -455,7 +455,9 @@ def run_cplex(
 
         objective = m.solution.get_objective_value()
 
-        solution = pd.Series(m.solution.get_values(), m.variables.get_names())
+        solution = pd.Series(
+            m.solution.get_values(), m.variables.get_names(), dtype=float
+        )
         solution = set_int_index(solution)
 
         if is_lp:
@@ -556,11 +558,11 @@ def run_gurobi(
     def get_solver_solution() -> Solution:
         objective = m.ObjVal
 
-        sol = pd.Series({v.VarName: v.x for v in m.getVars()})
+        sol = pd.Series({v.VarName: v.x for v in m.getVars()}, dtype=float)
         sol = set_int_index(sol)
 
         try:
-            dual = pd.Series({c.ConstrName: c.Pi for c in m.getConstrs()})
+            dual = pd.Series({c.ConstrName: c.Pi for c in m.getConstrs()}, dtype=float)
             dual = set_int_index(dual)
         except AttributeError:
             logger.warning("Dual values of MILP couldn't be parsed")
@@ -647,12 +649,12 @@ def run_xpress(
 
         var = [str(v) for v in m.getVariable()]
 
-        sol = pd.Series(m.getSolution(var), index=var)
+        sol = pd.Series(m.getSolution(var), index=var, dtype=float)
         sol = set_int_index(sol)
 
         try:
             dual = [str(d) for d in m.getConstraint()]
-            dual = pd.Series(m.getDual(dual), index=dual)
+            dual = pd.Series(m.getDual(dual), index=dual, dtype=float)
             dual = set_int_index(dual)
         except xpress.SolverError:
             logger.warning("Dual values of MILP couldn't be parsed")
