+++ test/test_linear_expression.py
@@ -60,9 +60,34 @@ def test_empty_linexpr(m):
     LinearExpression(None, m)
 
 
+def test_linexpr_with_wrong_data(m):
+    with pytest.raises(ValueError):
+        LinearExpression(1, m)
+
+    with pytest.raises(ValueError):
+        LinearExpression(xr.Dataset({"a": [1]}), m)
+
+    coeffs = xr.DataArray([1, 2], dims=["a"])
+    vars = xr.DataArray([1, 2], dims=["a"])
+    data = xr.Dataset({"coeffs": coeffs, "vars": vars})
+    with pytest.raises(ValueError):
+        LinearExpression(data, m)
+
+    coords = [pd.Index([0], name="a"), pd.Index([1, 2], name="_term")]
+    coeffs = xr.DataArray(np.array([[1, 2]]), coords=coords)
+    vars = xr.DataArray(np.array([[1, 2]]), coords=coords)
+    data = xr.Dataset({"coeffs": coeffs, "vars": vars})
+    with pytest.raises(ValueError):
+        LinearExpression(data, None)
+
+
 def test_repr(m):
     expr = m.linexpr((10, "x"), (1, "y"))
     expr.__repr__()
+
+
+def test_fill_value():
+    isinstance(LinearExpression.fill_value, dict)
 
 
 def test_linexpr_with_scalars(m):
@@ -138,7 +163,7 @@ def test_linear_expression_multi_indexed(u):
     assert isinstance(expr, LinearExpression)
 
 
-def test_linear_expression_with_errors(x):
+def test_linear_expression_with_errors(m, x):
     with pytest.raises(TypeError):
         x.to_linexpr(x)
 
@@ -160,17 +185,20 @@ def test_linear_expression_with_errors(x):
     with pytest.raises(TypeError):
         x / (1 * x)
 
+    with pytest.raises(TypeError):
+        m.linexpr((10, x.labels), (1, "y"))
+
 
 def test_linear_expression_from_rule(m, x, y):
     def bound(m, i):
         if i == 1:
-            return (i - 1) * x[i - 1] + y[i]
+            return (i - 1) * x[i - 1] + y[i] + 1 * x[i]
         else:
             return i * x[i] - y[i]
 
     expr = LinearExpression.from_rule(m, bound, x.coords)
     assert isinstance(expr, LinearExpression)
-    assert expr.nterm == 2
+    assert expr.nterm == 3
     repr(expr)  # test repr
 
 
