+++ linopy/variables.py
@@ -15,8 +15,8 @@ import dask
 import numpy as np
 import pandas as pd
 from deprecation import deprecated
-from numpy import arange, floating, hstack, inf, issubdtype
-from xarray import DataArray, Dataset, full_like, zeros_like
+from numpy import floating, inf, issubdtype
+from xarray import DataArray, Dataset, zeros_like
 from xarray.core import indexing, utils
 
 import linopy.expressions as expressions
