### 2023-01-10 af5c26f
constraints: fix init with pd.Series
