### 2023-01-11 637b7c0
constraints: fix scalar instance check on rhs examples: fix non-coord coeffs
