+++ doc/release_notes.rst
@@ -15,6 +15,7 @@ Upcoming Release
 * Coefficients of variables are no also allowed to be `np.nan`. These coefficients are ignored in the LP file writing.
 * The classes Variable, LinearExpression, Constraint, ScalarVariable, ScalarLinearExpression and ScalarConstraint now require the model in the initialization (mostly internal code is affected).
 * The `eval` module was removed in favor of arithmetic operations on the classes `Variable`, `LinearExpression` and `Constraint`.
+* Solver options are now printed out in the console when solving a model.
 
 Version 0.0.15
 --------------
