### 2023-01-12 fe497be
raise UserWarning when adding unaligned variables
