+++ test/test_io.py
@@ -20,7 +20,7 @@ def m():
     m = Model()
 
     x = m.add_variables(4, pd.Series([8, 10]), name="x")
-    y = m.add_variables(0, pd.DataFrame([[1, 2], [3, 4], [5, 6]]), name="y")
+    y = m.add_variables(0, pd.DataFrame([[1, 2], [3, 4]]), name="y")
 
     m.add_constraints(x + y, LESS_EQUAL, 10)
 
@@ -48,11 +48,14 @@ def test_str_arrays_with_nans():
     m = Model()
 
     m.add_variables(4, pd.Series([8, 10]), name="x")
-    # now expand the second dimension, expended values of x will be nan
-    m.add_variables(0, pd.DataFrame([[1, 2], [3, 4], [5, 6]]), name="y")
-    assert m["x"].values[-1] == -1
+    # now expand the second dimension, expanded values of x will be nan
 
-    da = int_to_str(m["x"].values)
+    with pytest.warns(UserWarning):
+        m.add_variables(0, pd.DataFrame([[1, 2]]), name="y")
+
+    assert m["y"].values[-1, -1] == -1
+
+    da = int_to_str(m["y"].values)
     assert da.dtype == object
 
 
