### 2023-01-12 3395520
fix single variable repr, consolidate repr code between Variable and Constraint
