### 2023-01-12 8412970
fix rename function enable automatic for coefficient assignment in from_tuple if needed
