### 2023-01-12 f727c8e
add test for missing coords
