### 2023-01-12 e16a804
linear expression: fix pandas object multiplication, fix large print of linear expressions
