### 2023-01-14 99af7b0
add config.py enable 'override' option if LinearExpression's are of the same shape
