+++ linopy/expressions.py
@@ -30,6 +30,7 @@ from linopy.common import (
     print_coord,
     print_single_expression,
 )
+from linopy.config import options
 from linopy.constants import EQUAL, GREATER_EQUAL, LESS_EQUAL
 
 
@@ -214,24 +215,33 @@ class LinearExpression:
             return f"LinearExpression:\n-----------------\n{expr_string}"
 
         # print only a few values
-        max_print = 14
+        max_print = options["display_max_rows"]
         split_at = max_print // 2
         to_print = head_tail_range(nexprs, max_print)
+        truncate = nexprs > max_print
         coords = self.unravel_coords(to_print)
 
         # loop over all values to print
-        data_string = ""
+        coord_strings = []
+        expr_strings = []
+        trunc_strings = []
         for i, coord in enumerate(coords):
 
-            coord_string = print_coord(coord)
+            coord_string = print_coord(coord) + ":"
             expr_string = print_single_expression(
                 self.coeffs.loc[coord].values, self.vars.loc[coord].values, self.model
             )
+            trunc_string = "\n\t\t..." if i == split_at - 1 and truncate else ""
 
-            data_string += f"\n{coord_string}:  {expr_string}"
+            coord_strings.append(coord_string)
+            expr_strings.append(expr_string)
+            trunc_strings.append(trunc_string)
 
-            if i == split_at - 1 and nexprs > max_print:
-                data_string += "\n\t\t..."
+        coord_width = max(len(c) for c in coord_strings)
+
+        data_string = ""
+        for c, e, t in zip(coord_strings, expr_strings, trunc_strings):
+            data_string += f"\n{c:<{coord_width}} {e}{t}"
 
         # create shape string
         nonterm_dims = [(k, v) for k, v in self.dims.items() if not k.startswith("_")]
@@ -250,6 +260,13 @@ class LinearExpression:
         if isinstance(other, variables.Variable):
             other = LinearExpression.from_tuples((1, other))
         return merge(self, other)
+
+    def __radd__(self, other):
+        # This is needed for using python's sum function
+        if other == 0:
+            return self
+        else:
+            raise NotImplementedError
 
     def __sub__(self, other):
         """
@@ -279,7 +296,7 @@ class LinearExpression:
             )
         if isinstance(other, (pd.Series, pd.DataFrame)):
             other = xr.DataArray(other)
-        coeffs = other * self.coeffs
+        coeffs = self.coeffs * other
         assert set(coeffs.shape) == set(self.coeffs.shape)
         return self.assign(coeffs=coeffs)
 
@@ -334,6 +351,10 @@ class LinearExpression:
     def dims(self):
         # do explicitly sort as in vars (same as in coeffs)
         return {k: self.data.dims[k] for k in self.vars.dims}
+
+    @property
+    def non_helper_dims(self):
+        return {k: self.data.dims[k] for k in self.dims if not k.startswith("_")}
 
     @property
     def vars(self):
@@ -883,11 +904,15 @@ def _pd_dataframe_wo_axes_names(df):
     return False
 
 
-def merge(*exprs, dim="_term", cls=LinearExpression):
+def merge(*exprs, dim="_term", cls=LinearExpression, **kwargs):
     """
     Merge multiple linear expression together.
 
     This function is a bit faster than summing over multiple linear expressions.
+    In case that the a list of LinearExpression with exactly the same shape,
+    the concatenation uses the coordinates of the first object as a basis
+    which overrides the coordinates of the consecutive objects.
+
 
     Parameters
     ----------
@@ -895,15 +920,21 @@ def merge(*exprs, dim="_term", cls=LinearExpression):
[…]
