### 2023-01-14 852f670
add test for options setter
