+++ linopy/variables.py
@@ -321,6 +321,13 @@ class Variable:
                 "unsupported operand type(s) for +: " f"{type(self)} and {type(other)}"
             )
 
+    def __radd__(self, other):
+        # This is needed for using python's sum function
+        if other == 0:
+            return self
+        else:
+            raise NotImplementedError
+
     def __sub__(self, other):
         """
         Subtract linear expressions or other variables from the variables.
@@ -1101,6 +1108,8 @@ class ScalarVariable:
         # This is needed for using python's sum function
         if other == 0:
             return self
+        else:
+            raise NotImplementedError
 
     def __sub__(self, other):
         return self.to_scalar_linexpr(1) - other
