### 2023-01-16 6499127
constants: fix solver available printout
