+++ linopy/constants.py
@@ -200,12 +200,15 @@ class Result:
     solver_model: Optional[Any] = None
 
     def __repr__(self) -> str:
+        solver_model_string = (
+            "not available" if self.solver_model is None else "available"
+        )
         return (
             f"Status: {self.status.status.value}\n"
             f"Termination condition: {self.status.termination_condition.value}\n"
             f"Solution: {len(self.solution.primal)} primals, {len(self.solution.dual)} duals\n"
             f"Objective: {self.solution.objective:.2e}\n"
-            f"Solver model: {'not' or self.solver_model} available\n"
+            f"Solver model: {solver_model_string}\n"
             f"Solver message: {self.status.legacy_status}"
         )
 
