### 2023-01-17 9562ce6
constraints.py: fix repr for single constraints wo coordinates
