### 2023-01-17 360536e
update release notes
