### 2023-01-18 61b61b9
make highs default solver
