+++ benchmark/scripts/run-pulp.py
@@ -1,0 +1,6 @@
+from benchmark_pulp import model
+
+n = int(snakemake.wildcards.N)
+solver = snakemake.wildcards.solver
+integerlabels = snakemake.params.integerlabels
+model(n, solver, integerlabels)
