### 2023-01-30 022c82e
add gurobipy restructure workflow
