+++ benchmark/scripts/run-linopy.py
@@ -10,5 +10,4 @@ from benchmark_linopy import model
 
 n = int(snakemake.wildcards.N)
 solver = snakemake.wildcards.solver
-integerlabels = snakemake.params.integerlabels
-model(n, solver, integerlabels)
+model(n, solver)
