+++ benchmark/scripts/run-pyomo.py
@@ -11,5 +11,4 @@ from benchmark_pyomo import model
 
 n = int(snakemake.wildcards.N)
 solver = snakemake.wildcards.solver
-integerlabels = snakemake.params.integerlabels
-model(n, solver, integerlabels)
+model(n, solver)
