+++ benchmark/scripts/run-gurobipy.py
@@ -1,4 +1,4 @@
-from benchmark_pulp import model
+from benchmark_gurobipy import model
 
 n = int(snakemake.wildcards.N)
 solver = snakemake.wildcards.solver
