+++ benchmark/scripts/common.py
@@ -2,7 +2,8 @@ import gc
 from time import time
 
 import pandas as pd
-from memory_profiler import memory_usage
+
+# from memory_profiler import memory_usage
 
 
 def profile(nrange, func, *args):
@@ -16,9 +17,9 @@ def profile(nrange, func, *args):
         end = time()
         duration = end - start
 
-        memory = memory_usage((func, (N, *args)))
+        # memory = memory_usage((func, (N, *args)))
 
-        res.loc[N] = duration, max(memory)
+        res.loc[N] = duration  # , max(memory)
 
         gc.collect()
 
