+++ .pre-commit-config.yaml
@@ -10,7 +10,7 @@ repos:
     hooks:
     -   id: black
 -   repo: https://github.com/PyCQA/docformatter
-    rev: v1.5.1
+    rev: v1.6.0.rc1
     hooks:
     -   id: docformatter
         args: [--in-place, --make-summary-multi-line, --pre-summary-newline]
@@ -23,7 +23,7 @@ repos:
     hooks:
     -   id: absolufy-imports
 -   repo: https://github.com/PyCQA/isort
-    rev: 5.11.4
+    rev: 5.12.0
     hooks:
       - id: isort
         args: ["--profile", "black", "--filter-files"]
