+++ paper/paper.bib
@@ -1,4 +1,30 @@
 
+@manual{mosek,
+author = "MOSEK ApS",
+title = "MOSEK Optimizer API for Python 9.3.22",
+year = 2019,
+url = "https://docs.mosek.com/9.3/pythonapi/index.html#project"
+}
+
+@article{diamond2016cvxpy,
+author  = {Steven Diamond and Stephen Boyd},
+title   = {{CVXPY}: {A} {P}ython-embedded modeling language for convex optimization},
+journal = {Journal of Machine Learning Research},
+year    = {2016},
+volume  = {17},
+number  = {83},
+pages   = {1--5},
+}
+
+@article{agrawal2018rewriting,
+  author  = {Agrawal, Akshay and Verschueren, Robin and Diamond, Steven and Boyd, Stephen},
+  title   = {A rewriting system for convex optimization problems},
+  journal = {Journal of Control and Decision},
+  year    = {2018},
+  volume  = {5},
+  number  = {1},
+  pages   = {42--60},
+}
 @misc{amploptimizationinc.AMPLStramlinedModeling,
   title = {{{AMPL}} - {{Stramlined}} Modeling for Real Optimization},
   author = {AMPL Optimization Inc.},
