+++ paper/paper.md
@@ -32,7 +32,7 @@ Linopy is an open-source package written in Python to build and process linear a
 
 Decades after its inception [@dantzigLinearProgrammingExtensions1963a], mathematical optimization is nowadays of immense importance for business, industry  and governmental decision-making. Optimization is used to address various sorts of complex problems such as challenges of climate change, energy transitions or food supply. Typically, an optimization problem, i.e. a mathematical program, consists of one objective function to be numerically minimized and a set of constraints that restrict the underlying variables to external conditions. Algebraic Modeling Languages (AML) aim at facilitating mathematical programming by allowing the user to formulate large scale, complex problems with a high-level syntax similar to the mathematical notation. The formulated problem is then passed to the solver of choice where a solution is calculated. AMLs provide the most user-friendly interface possible to various solvers, each with its own set of features.
 
-Well established AMLs such as GAMS [@bussieckGeneralAlgebraicModeling2004] and AMPL [@fourerModelingLanguageMathematical1990] support a wide range of solvers, but are license-restricted and rely on closed-source code. In contrast, AMLs as JuMP [@dunningJuMPModelingLanguage2017], Pyomo [@hartPyomoOptimizationModeling2017], GEKKO [@bealGEKKOOptimizationSuite2018] and [PuLP](https://github.com/coin-or/pulp) are open-source and have gained increasing attention throughout the recent years. While the Julia package JuMP is characterized by high-performance, in-memory communication with the solvers, the Python packages Pyomo, GEKKO and PuLP lack parallelized, low-level operations and communicate slower with the solver via intermediate files written to disk. Further, the Python AMLs do not make use of state-of-the-art data handling packages. In particular, the assignment of coordinates or indexes is often not supported or memory extensive due to use of an object-oriented implementation where every single combination of coordinates is stored separately.
+Well established AMLs such as GAMS [@bussieckGeneralAlgebraicModeling2004] and AMPL [@fourerModelingLanguageMathematical1990] support a wide range of solvers, but are license-restricted and rely on closed-source code. In contrast, AMLs as JuMP [@dunningJuMPModelingLanguage2017], CVXPY [@agrawal2018rewriting], Pyomo [@hartPyomoOptimizationModeling2017], GEKKO [@bealGEKKOOptimizationSuite2018] and [PuLP](https://github.com/coin-or/pulp) are open-source and have gained increasing attention throughout the recent years. While the Julia package JuMP is characterized by high-performance, in-memory communication with the solvers, the Python packages Pyomo, GEKKO and PuLP lack parallelized, low-level operations and communicate slower with the solver via intermediate files written to disk. An exception is CVXPY, which supports fast array-based operations and uses low-level wrappers to the solvers. However, it is common among Python AMLs not to make use of state-of-the-art data handling packages. In particular, the assignment of coordinates or indexes is often not supported or memory extensive due to use of an object-oriented implementation where every single combination of coordinates is stored separately.
 
 Linopy is a open-source Python packages representing a new kind of AML that tackles these issues together. By introducing an array-based data model for variables and constraints, Linopy makes mathematical programming compatible with Python's advanced data handling packages Numpy [@harrisArrayProgrammingNumPy2020], Pandas [@rebackPandasdevPandasPandas2022] and Xarray [@hoyerXarrayNDLabeled2017].
 
@@ -40,7 +40,7 @@ The approach follows the idea that a variable $x(d_1, d_2, ..., d_K)$ may be def
 
[…]
