+++ benchmark/scripts/merge-benchmarks.py
@@ -19,17 +19,22 @@ df["Number of Constraints"] = df.N**2 * 2
 solver_memory = df.loc[df.API == "Solving Process", "Memory"].values
 solver_time = df.loc[df.API == "Solving Process", "Time"].values
 
+# Make a correction of the memory usage, some APIs use external processes for the solving process
+api_with_external_process = {"pyomo"}
+api_with_internal_process = set(snakemake.params.apis).difference(
+    api_with_external_process
+)
+
+
 absolute = df.copy()
-# Make a correction of the memory usage:
-# Pyomo uses external threads for the solving process, this is not counted by the snakemake
-# memory tracking
-absolute.loc[absolute.API == "pyomo", "Memory"] += solver_memory
+for api in api_with_external_process:
+    absolute.loc[absolute.API == api, "Memory"] += solver_memory
 absolute.to_csv(snakemake.output.absolute)
 
 overhead = df.copy()
-for api in ["jump", "pyomo", "linopy", "pulp"]:
+for api in snakemake.params.apis:
     overhead.loc[overhead.API == api, "Time"] -= solver_time
-overhead.loc[overhead.API == "jump", "Memory"] -= solver_memory
-overhead.loc[overhead.API == "linopy", "Memory"] -= solver_memory
+for api in api_with_internal_process:
+    overhead.loc[overhead.API == api, "Memory"] -= solver_memory
 overhead = overhead.query("API != 'Solving Process'")
 overhead.to_csv(snakemake.output.overhead)
