+++ benchmark/scripts/run-cvxpy.py
@@ -1,0 +1,5 @@
+from benchmark_gurobipy import model
+
+n = int(snakemake.wildcards.N)
+solver = snakemake.wildcards.solver
+model(n, solver)
