+++ benchmark/scripts/benchmark_linopy.py
@@ -7,9 +7,8 @@ Created on Fri Nov 19 17:40:33 2021.
 """
 
 
-import pandas as pd
 from common import profile
-from numpy import arange, cos, sin
+from numpy import arange
 
 from linopy import Model
 
