+++ benchmark/environment.yaml
@@ -16,6 +16,7 @@ dependencies:
 - snakemake
 - memory_profiler
 - jupyter
+- cvxpy
 
 - pip:
   - ortools>=9.5
