### 2023-02-02 PyPSA/linopy#95 @FabianHofmann
Add more complex example to the documentation
To showcase the full set of flexibility given by the data-handling through xarray, it would be good to create more complex examples. These should include operations like:

- selection (sel / drop_sel)
- where / masking
- broadcast_like
- diff
- shift
- fill (bfill/ffill)
- groupby
- rolling
- operations on different axis
