+++ benchmark/scripts/run-pulp.py
@@ -1,4 +1,9 @@
-from benchmark_pulp import model
+from benchmark_pulp import basic_model, knapsack_model
+
+if snakemake.config["benchmark"] == "basic":
+    model = basic_model
+elif snakemake.config["benchmark"] == "knapsack":
+    model = knapsack_model
 
 n = int(snakemake.wildcards.N)
 solver = snakemake.wildcards.solver
