+++ benchmark/scripts/benchmark_linopy.py
@@ -9,11 +9,15 @@ Created on Fri Nov 19 17:40:33 2021.
 
 from common import profile
 from numpy import arange
+from numpy.random import randint, seed
 
 from linopy import Model
 
+# Random seed for reproducibility
+seed(125)
 
-def model(n, solver):
+
+def basic_model(n, solver):
     m = Model()
     N, M = [arange(n), arange(n)]
     x = m.add_variables(coords=[N, M])
@@ -23,11 +27,27 @@ def model(n, solver):
     m.add_objective((2 * x).sum() + y.sum())
     # m.to_file(f"linopy-model.lp")
     m.solve(solver)
-    return
+    return m.objective_value
+
+
+def knapsack_model(n, solver):
+    m = Model()
+    packages = m.add_variables(coords=[arange(n)], binary=True)
+    weight = randint(1, 100, size=n)
+    value = randint(1, 100, size=n)
+    m.add_constraints((weight * packages).sum() <= 200)
+    m.add_objective(-(value * packages).sum())  # use minus because of minimization
+    m.solve(solver_name=solver)
+    return -m.objective_value
 
 
 if __name__ == "__main__":
     solver = snakemake.wildcards.solver
+
+    if snakemake.config["benchmark"] == "basic":
+        model = basic_model
+    elif snakemake.config["benchmark"] == "knapsack":
+        model = knapsack_model
 
     # dry run first
     model(2, solver)
