+++ benchmark/scripts/benchmark_jump.jl
@@ -3,8 +3,10 @@ using Gurobi
 using DataFrames
 using CSV
 using Dates
+using Random
+Random.seed!(125)
 
-function model(n, solver)
+function basic_model(n, solver)
     m = Model(Gurobi.Optimizer)
     N = 1:n
     M = 1:n
@@ -14,8 +16,21 @@ function model(n, solver)
     @constraint(m, [i=N, j=M], x[i, j] + y[i, j] >= 0)
     @objective(m, Min, sum(2 * x[i, j] + y[i, j] for i in N, j in M))
     optimize!(m)
-    return m
+    return objective_value(m)
 end
+
+function knapsack_model(n, solver)
+    m = Model(solver)
+    N = 1:n
+    @variable(m, x[N], Bin)
+    weight = rand(1:100, n)
+    value = rand(1:100, n)
+    @constraint(m, sum(weight[i] * x[i] for i in N) <= 200)
+    @objective(m, Max, sum(value[i] * x[i] for i in N))
+    optimize!(m)
+    return objective_value(m)
+end
+
 
 if snakemake.wildcards["solver"] == "gurobi"
     solver = Gurobi.Optimizer
@@ -24,15 +39,22 @@ elseif snakemake.wildcards["solver"] == "cbc"
     solver = CBC.Optimizer
 end
 
+if snakemake.config["benchmark"] == "basic"
+    model = basic_model
+elseif snakemake.config["benchmark"] == "knapsack"
+    model = knapsack_model
+end
+
 # jit compile everything
 model(1, solver)
 
-profile = DataFrame(N=Int[], Time=Float64[], Memory=Float64[])
+profile = DataFrame(N=Int[], Time=Float64[], Memory=Float64[], Objective=Float64[])
 
 for N in snakemake.params[1]
     mem = @allocated(model(N, solver))/10^6
     time = @elapsed(model(N, solver))
-    push!(profile, [N, time, mem])
+    objective = model(N, solver)
+    push!(profile, [N, time, mem, objective])
 end
 profile[!, :API] .= "jump"
 insertcols!(profile, 1, :Row => 1:nrow(profile))
