+++ benchmark/scripts/benchmark_cvxpy.py
@@ -1,9 +1,13 @@
 import cvxpy as cp
 import numpy as np
 from common import profile
+from numpy.random import randint, seed
+
+# Random seed for reproducibility
+seed(125)
 
 
-def model(n, solver):
+def basic_model(n, solver):
     # Create variables
     x = cp.Variable((n, n))
     y = cp.Variable((n, n))
@@ -17,11 +21,37 @@ def model(n, solver):
     m = cp.Problem(objective, constraints)
     m.solve(solver=solver.upper())
 
-    return m
+    return m.value
+
+
+def knapsack_model(n, solver):
+    # Define the variables
+    weight = randint(1, 100, size=n)
+    value = randint(1, 100, size=n)
+
+    x = cp.Variable(n, boolean=True)
+
+    # Define the constraints
+    constraints = [weight @ x <= 200]
+
+    # Define the objective function
+    objective = cp.Maximize(value @ x)
+
+    # Optimize the model
+    m = cp.Problem(objective, constraints)
+    m.solve(solver=solver.upper())
+
+    # return objective
+    return m.value
 
 
 if __name__ == "__main__":
     solver = snakemake.wildcards.solver
+
+    if snakemake.config["benchmark"] == "basic":
+        model = basic_model
+    elif snakemake.config["benchmark"] == "knapsack":
+        model = knapsack_model
 
     # dry run first
     model(2, solver)
