+++ benchmark/scripts/benchmark_pyomo.py
@@ -76,7 +76,7 @@ def knapsack_model(n, solver):
 
 
 if __name__ == "__main__":
-    solver = snakemake.wildcards.solver
+    solver = snakemake.config["solver"]
 
     if snakemake.config["benchmark"] == "basic":
         model = basic_model
