+++ benchmark/config.yaml
@@ -1,14 +1,14 @@
 
-benchmark: 'basic'
+benchmark: 'knapsack'
+
+solver: gurobi
 
 basic:
   name: basic
   nrange: [20, 50, 100, 180, 400, 600, 900, 1200]
-  solver: gurobi
   apis: ['linopy', 'jump', 'pyomo', 'cvxpy', 'pulp', 'gurobipy', 'solver']
 
 knapsack:
   name: knapsack
-  nrange: [10, 50, 100]
-  solver: gurobi
+  nrange: [1000, 100000, 500000, 1000000, 3000000]
   apis: ['linopy', 'jump', 'pyomo', 'cvxpy', 'pulp', 'gurobipy', 'solver']
