+++ CONTRIBUTING.md
@@ -1,0 +1,1 @@
+Linopy's contributor guidelines can be found in the official [documentation](https://linopy.readthedocs.io/en/master/contributing.html).
