+++ benchmark/scripts/concat-benchmarks.py
@@ -6,20 +6,35 @@ Created on Wed Feb  9 09:34:00 2022.
 @author: fabian
 """
 
+from pathlib import Path
+
 import pandas as pd
 
-dfs = [pd.read_csv(fn, sep="\t") for fn in snakemake.input]
-dfs = pd.concat(dfs, axis=0, ignore_index=True)
-dfs["N"] = snakemake.params.nrange
+dfs = [pd.read_csv(fn, sep="\t") for fn in snakemake.input.memory]
+df = pd.concat(dfs, axis=0, ignore_index=True)
+df["N"] = snakemake.params.nrange
 
-dfs = dfs.rename(columns={"s": "Time", "max_rss": "Memory"})
-dfs = dfs.replace("-", 0)
-dfs["Memory"] = dfs["Memory"].astype(float)
+df = df.rename(columns={"s": "Time", "max_rss": "Memory"})
+df = df.replace("-", 0)
+df["Memory"] = df["Memory"].astype(float)
 
-if snakemake.wildcards.api == "solver":
-    dfs["API"] = "Solving Process"
+if snakemake.params.api == "solver":
+    df["API"] = "Solving Process"
 else:
-    dfs["API"] = snakemake.wildcards.api
+    df["API"] = snakemake.params.api
 
-cols = ["N", "Time", "Memory", "API"]
-dfs[cols].to_csv(snakemake.output[0])
+df = df[["N", "Time", "Memory", "API"]]
+
+benchmark_time = snakemake.input.time
+if benchmark_time is not None:
+
+    if isinstance(benchmark_time, str):
+        df_time = pd.read_csv(benchmark_time, index_col=0)
+        df["Time"] = df_time.Time.values
+        df["Objective"] = df_time.Objective.values
+    else:
+        # for solvers we need to read the time from the single output files
+        df["Time"] = [float(Path(fn).read_text()) for fn in benchmark_time]
+
+
+df.to_csv(snakemake.output.benchmark)
