+++ benchmark/scripts/leftovers/benchmarks-pypsa-eur/common.py
@@ -1,0 +1,11 @@
+PATH = "/home/fabian/vres/py/pypsa-eur/networks/elec_s_37.nc"
+NSNAPSHOTS = -1
+SOLVER = "gurobi"
+SOLVER_PARAMS = {
+    "crossover": 0,
+    "method": 2,
+    "BarConvTol": 1.0e-3,
+    "FeasibilityTol": 1.0e-3,
+}
+# SOLVER = "cbc"
+# SOLVER_PARAMS = {}
