+++ paper/paper.md
@@ -40,7 +40,7 @@ The approach follows the idea that a variable $x(d_1, d_2, ..., d_K)$ may be def
 
 The array-based modelling approach does not only lead to more flexibility but also increases the overall performance. The following figure shows the [benchmark](https://github.com/PyPSA/linopy/tree/master/benchmark) against the AMLs JuMP and Pyomo. The two packages are are both open source and well-established and therefore suitable for the comparison. Linopy outperforms Pyomo in terms of speed and memory-efficiency and outperforms JuMP in memory-efficiency.
 
-![Benchmark of Linopy against JuMP and Pyomo with the producing Snakemake workflow available [here](https://github.com/PyPSA/linopy/tree/master/benchmark). The benchmark is based on a simple linear problem and uses the Gurobi solver. The overhead is calculated from the difference of the whole solving process via the AML and the solving process on the solver side alone. Note that the benchmark is hardly dependent on the complexity of the problem. Thus, adding more terms to the constraints, setting different kind of index labels or changing it to an mixed-integer problem does hardly have an effect on the overhead. \label{fig:benchmark}](figures/benchmark-overhead.pdf)
+![Benchmark of Linopy against comparable packages with the producing Snakemake workflow available [here](https://github.com/PyPSA/linopy/tree/master/benchmark). The benchmark is based on a 1-dimensional knapsack problem and uses the Gurobi solver. The overhead is calculated from the difference of the whole solving process via the AML and the solving process on the solver side alone. Note that the benchmark is hardly dependent on the complexity of the problem. Thus, adding more terms to the constraints, setting different kind of index labels or changing it to an mixed-integer problem does hardly have an effect on the overhead. \label{fig:benchmark}](figures/benchmark-overhead.pdf)
 
 Due to a strong alignment to the Xarray package, Linopy supports storing the optimization model as a NetCDF file [@rewNetCDFInterfaceScientific1990], which allows users to quickly share optimization problems with others. Using the [Paramiko package](https://paramiko.org), Linopy offers the user to send unsolved problems to a server and retrieve the solution after running the optimization remotely, which is particularly helpful if large computing resources are needed.
 
