+++ paper/paper.bib
@@ -1,30 +1,13 @@
-
-@manual{mosek,
-author = "MOSEK ApS",
-title = "MOSEK Optimizer API for Python 9.3.22",
-year = 2019,
-url = "https://docs.mosek.com/9.3/pythonapi/index.html#project"
-}
-
-@article{diamond2016cvxpy,
-author  = {Steven Diamond and Stephen Boyd},
-title   = {{CVXPY}: {A} {P}ython-embedded modeling language for convex optimization},
-journal = {Journal of Machine Learning Research},
-year    = {2016},
-volume  = {17},
-number  = {83},
-pages   = {1--5},
-}
-
-@article{agrawal2018rewriting,
-  author  = {Agrawal, Akshay and Verschueren, Robin and Diamond, Steven and Boyd, Stephen},
-  title   = {A rewriting system for convex optimization problems},
+@article{agrawalRewritingSystemConvex2018,
+  title = {A Rewriting System for Convex Optimization Problems},
+  author = {Agrawal, Akshay and Verschueren, Robin and Diamond, Steven and Boyd, Stephen},
+  year = {2018},
   journal = {Journal of Control and Decision},
-  year    = {2018},
-  volume  = {5},
-  number  = {1},
-  pages   = {42--60},
+  volume = {5},
+  number = {1},
+  pages = {42--60}
 }
+
 @misc{amploptimizationinc.AMPLStramlinedModeling,
   title = {{{AMPL}} - {{Stramlined}} Modeling for Real Optimization},
   author = {AMPL Optimization Inc.},
@@ -41,6 +24,12 @@ pages   = {1--5},
   journal = {Anaconda Documentation},
   publisher = {{Anaconda Inc.}},
   version = {Vers. 2-2.4.0}
+}
+
+@book{apsMOSEKOptimizerAPI2019,
+  title = {{{MOSEK Optimizer API}} for {{Python}} 9.3.22},
+  author = {ApS, MOSEK},
+  year = {2019}
 }
 
 @article{bealGEKKOOptimizationSuite2018,
@@ -172,6 +161,16 @@ pages   = {1--5},
   title = {Dask: {{Library}} for Dynamic Task Scheduling},
   author = {{Dask Development Team}},
   year = {2016}
+}
+
+@article{diamondCVXPYPythonembeddedModeling2016,
+  title = {{{CVXPY}}: {{A Python-embedded}} Modeling Language for Convex Optimization},
+  author = {Diamond, Steven and Boyd, Stephen},
+  year = {2016},
+  journal = {Journal of Machine Learning Research},
+  volume = {17},
+  number = {83},
+  pages = {1--5}
 }
 
 @article{dunningJuMPModelingLanguage2017,
@@ -380,6 +379,16 @@ pages   = {1--5},
   file = {/home/fabian/snap/zotero-snap/common/Zotero/storage/M5UU8G7J/List_of_optimization_software.html}
 }
 
+@misc{mitchelPulp2022,
+  title = {Pulp},
+  author = {Mitchel, Stuart and Kean, Anita and Mason, Andrew and O'Sullivan, Michael and Phillips, Antony and Peschiera, Franco},
+  year = {2022},
+  month = jul,
+  abstract = {A python Linear Programming API},
+  howpublished = {COIN-OR Foundation},
+  keywords = {constraints,mip,pulp,python,solver}
+}
+
 @article{morshedGeneralizedAffineScaling2020,
   title = {Generalized Affine Scaling Algorithms for Linear Programming Problems},
   author = {Morshed, Md Sarowar and {Noor-E-Alam}, Md.},
@@ -394,6 +403,12 @@ pages   = {1--5},
[…]
