### 2023-02-02 #4823 @FabianHofmann
Hey @toressa,

first of all thank you for your detailed review! These are all very helpful comments and I could finally find some time to address them. Since version 0.0.11 actually a lot has happened. A lot has to do with stabilizing the API and the data handling.
I will try to address all your comments in the following.

# Benchmarks

The benchmark was extended. It now includes the following models:

- knapsack
- the original model

which can chosen by the user. The figure in the paper shows the benchmark of the knapsack problem. It compares the following apis:

- linopy
- JuMP
- cvxpy
- pulp
- pyomo
- gurobipy (matrix api)
- ortools

However, ortools does not yet support gurobi v10, therefore it did not make it into the final figure. The knapsack model and the original problem lead to very similar overheads. So I hope it is okay to not extend the set of benchmark models further. As far I see it, the solution retrieval is included by the all AML's.

# General code

I added a `constants` module to define sets of constant values used in linopy, i.e. senses, solver status, termination condition. It also includes a solution class definition and result class definition (used for consistent print out of results).

To answer your question about the preferred tool, from the computational point of view, it does not make a difference which data handling tool (numpy, pandas, xarray) is used, as they all rely on numpy arrays in the background. However, I pointed out in the [doc](https://linopy.readthedocs.io/en/latest/creating-variables.html#Using-xarray-DataArray's) that xarray is the most convenient and safe tool to create variables.

# General

I added a `CONTRIBUTING.md` file to the repository. It points to the [contributing guidelines](https://linopy.readthedocs.io/en/latest/contributing.html) in the documentation.

As for the PWL and QC, we discussed a rough but promising structure how to represent them. But for the proper implementation, I will need to have some spare time, or have a motivated and competent student to work on it. So I hope it is fine to not make a requirement for the JOSS contribution.

Good point about the more complex example cases. There were already some users asking for it. I created an [issue](https://github.com/PyPSA/linopy/issues/95) that I will work on in the next weeks.

# File Types

NetCDF is a data format that is used in many scientific fields. It is a very convenient format to store multidimensional data. It is also the default format for xarray, is extremely fast, and can be used to store large amounts of data. In OR, NetCDF is not used to my knowledge. Its use case here is not transferring a linopy model to other APIs but rather store a linopy model to fast disk. It is also used when solving a linopy model on a remote server.

The export in MPS file format is now also supported by linopy. And tested across all solvers https://github.com/PyPSA/linopy/pull/97

The executable files were corrected.

# Citation

I added a citation for PuLP

About the naming. Fair point. However I wanted to build the bridge between the data science and the OR community. So, unless I have a brainwave, I would stick to that name.

---

I hope that tackles all the point so far!

Thanks again and looking forward to hear from you!

Also pinging @g4brielvs whether there are remaining things on the list that I should cover for the package :)
