### 2023-02-06 17806bb
solvers: fix dual retrieval from highs
