+++ linopy/common.py
@@ -151,6 +151,7 @@ def has_optimized_model(func):
     """
     Check if a reference model is set.
     """
+
     @wraps(func)
     def wrapper(self, *args, **kwargs):
         if self.model is None:
@@ -165,6 +166,7 @@ def has_optimized_model(func):
 def is_constant(func):
     from linopy import expressions, variables
 
+    #
     @wraps(func)
     def wrapper(self, arg):
         if isinstance(arg, (variables.Variable, expressions.LinearExpression)):
@@ -175,6 +177,7 @@ def is_constant(func):
 
 
 def forward_as_properties(**routes):
+    #
     def add_accessor(cls, item, attr):
         @property
         def get(self):
