+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
     -   id: trailing-whitespace
     -   id: check-merge-conflict
 -   repo: https://github.com/psf/black
-    rev: 22.12.0
+    rev: 23.1.0
     hooks:
     -   id: black
     # Current docformatter version has a problem with urls (it's annoying)
