### 2023-03-01 c73468e
added test, LinExpr constructor, __add__
