### 2023-03-03 70b6b6b
updates, split tests, rewrote radd
