### 2023-03-06 #4823 @torressa
Thanks for all the work!

I think I am happy now with the code.
Just a few more items for the paper itself.
I ran the benchmarks and I think the experiments in the paper need a bit more
detail. Particularly, version numbers (e.g. Python, gurobipy, Julia, JuMP) and machine run and benchmark model (i.e. knapsack or other).

When I ran these on my machine the plots looked slightly different that's why I am asking.
(I also had to modify the profiling function to populate the memory field for it to produce a plot).

Citations:
- Gurobi: [How do I cite Gurobi software for an academic publication?](https://support.gurobi.com/hc/en-us/articles/360013195592-How-do-I-cite-Gurobi-software-for-an-academic-publication-)

```bibtex
@misc{gurobi,
  author = {{Gurobi Optimization, LLC}},
  title = {{Gurobi Optimizer Reference Manual}},
  year = 2023,
  url = "https://www.gurobi.com"
}
```

I didn't really find citation details for the other solvers.
