+++ paper/paper.bib
@@ -1,506 +1,504 @@
 @article{agrawalRewritingSystemConvex2018,
-  title = {A Rewriting System for Convex Optimization Problems},
-  author = {Agrawal, Akshay and Verschueren, Robin and Diamond, Steven and Boyd, Stephen},
-  year = {2018},
+  title   = {A Rewriting System for Convex Optimization Problems},
+  author  = {Agrawal, Akshay and Verschueren, Robin and Diamond, Steven and Boyd, Stephen},
+  year    = {2018},
   journal = {Journal of Control and Decision},
-  volume = {5},
-  number = {1},
-  pages = {42--60}
+  volume  = {5},
+  number  = {1},
+  pages   = {42--60}
 }
 
 @misc{amploptimizationinc.AMPLStramlinedModeling,
-  title = {{{AMPL}} - {{Stramlined}} Modeling for Real Optimization},
-  author = {AMPL Optimization Inc.},
-  journal = {AMPL},
-  abstract = {Why AMPL? The AMPL system supports the entire optimization modeling lifecycle \textemdash{} formulation, testing, deployment, and maintenance \textemdash{} in an integrated way that promotes rapid development and reliable results. Using a high-level algebraic representation that describes optimization models in the...},
+  title        = {{{AMPL}} - {{Stramlined}} Modeling for Real Optimization},
+  author       = {AMPL Optimization Inc.},
+  journal      = {AMPL},
+  abstract     = {Why AMPL? The AMPL system supports the entire optimization modeling lifecycle \textemdash{} formulation, testing, deployment, and maintenance \textemdash{} in an integrated way that promotes rapid development and reliable results. Using a high-level algebraic representation that describes optimization models in the...},
   howpublished = {https://ampl.com/},
-  langid = {american},
-  file = {/home/fabian/snap/zotero-snap/common/Zotero/storage/JNJKCUAT/ampl.com.html}
+  langid       = {american},
+  file         = {/home/fabian/snap/zotero-snap/common/Zotero/storage/JNJKCUAT/ampl.com.html}
 }
 
 @article{AnacondaSoftwareDistribution2020,
-  title = {Anaconda Software Distribution},
-  year = {2020},
-  journal = {Anaconda Documentation},
+  title     = {Anaconda Software Distribution},
+  year      = {2020},
+  journal   = {Anaconda Documentation},
   publisher = {{Anaconda Inc.}},
-  version = {Vers. 2-2.4.0}
+  version   = {Vers. 2-2.4.0}
 }
 
 @book{apsMOSEKOptimizerAPI2019,
-  title = {{{MOSEK Optimizer API}} for {{Python}} 9.3.22},
+  title  = {{{MOSEK Optimizer API}} for {{Python}} 9.3.22},
   author = {ApS, MOSEK},
-  year = {2019}
+  year   = {2019}
 }
 
 @article{bealGEKKOOptimizationSuite2018,
-  title = {{{GEKKO Optimization Suite}}},
-  author = {Beal, Logan and Hill, Daniel and Martin, R. and Hedengren, John},
-  year = {2018},
-  month = jul,
-  journal = {Processes},
-  volume = {6},
-  number = {8},
-  pages = {106},
-  issn = {2227-9717},
-  doi = {10.3390/pr6080106},
+  title    = {{{GEKKO Optimization Suite}}},
+  author   = {Beal, Logan and Hill, Daniel and Martin, R. and Hedengren, John},
+  year     = {2018},
+  month    = jul,
+  journal  = {Processes},
+  volume   = {6},
+  number   = {8},
+  pages    = {106},
+  issn     = {2227-9717},
+  doi      = {10.3390/pr6080106},
[…]
