+++ benchmark/README.md
@@ -1,15 +1,55 @@
-# Benchmark Workflow
+# Benchmark
+
+![Resources benchmark](benchmark_resource-overhead.pdf)
+
+This benchmark compares the performance of `linopy` against similar packages considering memory and time overhead for solving an optimization problem with N variable. The overhead is defined by the resources needed to define the problem, pass it to the solver and retrieve the solution. For reproduce the benchmark, this directory contains all necessary code in a Snakemake workflow. The workflow allows to choose between two different optimization problems to be solved
+
+1. A one-dimensional knapsack problem, which is a simple linear program with the following formulation described [here](https://www.wikiwand.com/en/Knapsack_problem) and used in the figure above. The problem is solved for a range of different values `N` representing the number of packages.  
+
+2. The second problem choice `"basic"` is a simple linear program with the following formulation
+    <p><span class="math display">∑<sub><em>i</em>, <em>j</em></sub>2<em>x</em><sub><em>i</em>, <em>j</em></sub> + <em>y</em><sub><em>i</em>, <em>j</em></sub></span></p>
+
+    s.t.
+
+    <span class="math display"><em>x</em><sub><em>i</em>, <em>j</em></sub> − <em>y</em><sub><em>i</em>, <em>j</em></sub> ≥ <em>i</em>   ∀ <em>i</em>, <em>j</em> ∈ {1, ..., <em>N</em>}</span>
+
+    <span class="math display"><em>x</em><sub><em>i</em>, <em>j</em></sub> + <em>y</em><sub><em>i</em>, <em>j</em></sub> ≥ 0   ∀ <em>i</em>, <em>j</em> ∈ {1, ..., <em>N</em>}</span></p>
+
+    which is initialized and solved for different values of `N` with each of the API's.
 
 
-This directory contains a Snakemake workflow benchmarking the performance of `linopy` against `pyomo` and `JuMP`. The linear problem
+To run the benchmark, install the conda environment with
 
+```bash
+conda env create -f environment.yaml
+conda activate linopy-benchmark
+```
 
-<p><span class="math display">∑<sub><em>i</em>, <em>j</em></sub>2<em>x</em><sub><em>i</em>, <em>j</em></sub> + <em>y</em><sub><em>i</em>, <em>j</em></sub></span></p>
+Replace `environment.yaml` by `environment.fixed.yaml` if you want to use the fixed versions of the packages used for creating the figure above and specified below.
+Make sure to have a working installation of `Julia` and `JuMP`, and optionally install the version stated below.
 
-s.t.
+Then, run the benchmark with
 
-<span class="math display"><em>x</em><sub><em>i</em>, <em>j</em></sub> − <em>y</em><sub><em>i</em>, <em>j</em></sub> ≥ <em>i</em>   ∀ <em>i</em>, <em>j</em> ∈ {1, ..., <em>N</em>}</span>
+```bash
+snakemake --cores 4
+```
 
-<span class="math display"><em>x</em><sub><em>i</em>, <em>j</em></sub> + <em>y</em><sub><em>i</em>, <em>j</em></sub> ≥ 0   ∀ <em>i</em>, <em>j</em> ∈ {1, ..., <em>N</em>}</span></p>
+### Versions Specfications
 
-is initialized and solved for different values of `N` with each of the API's.
+For creating the figure above, the following versions of the packages were used
+
+- python v3.9.0
+- cxvpy v1.3
+- gurobipy v10.0
+- pulp v2.7
+- Pyomo v6.4.4
+- ortools v9.5
+- linopy v0.1.3
+- julia v1.6.7
+- JuMP v1.1.1
+
[…]
