+++ benchmark/environment.fixed.yaml
@@ -1,0 +1,340 @@
+name: linopy-benchmark
+channels:
+  - bioconda
+  - gurobi
+  - http://conda.anaconda.org/gurobi
+  - defaults
+  - conda-forge
+dependencies:
+  - _libgcc_mutex=0.1
+  - _openmp_mutex=4.5
+  - absl-py=1.4.0
+  - aioeasywebdav=2.4.0
+  - aiohttp=3.8.3
+  - aiosignal=1.3.1
+  - alsa-lib=1.2.3.2
+  - amply=0.1.5
+  - anyio=3.6.2
+  - appdirs=1.4.4
+  - argon2-cffi=21.3.0
+  - argon2-cffi-bindings=21.2.0
+  - asttokens=2.2.1
+  - async-timeout=4.0.2
+  - attmap=0.13.2
+  - attrs=22.2.0
+  - backcall=0.2.0
+  - backports=1.0
+  - backports.functools_lru_cache=1.6.4
+  - bcrypt=3.2.2
+  - beautifulsoup4=4.11.1
+  - bleach=6.0.0
+  - boto3=1.26.58
+  - botocore=1.29.58
+  - bottleneck=1.3.6
+  - brotli=1.0.9
+  - brotli-bin=1.0.9
+  - brotlipy=0.7.0
+  - bzip2=1.0.8
+  - c-ares=1.18.1
+  - ca-certificates=2023.01.10
+  - cachetools=5.3.0
+  - certifi=2022.12.7
+  - cffi=1.14.6
+  - charset-normalizer=2.1.1
+  - click=8.1.3
+  - cloudpickle=2.2.1
+  - coin-or-cbc=2.10.8
+  - coin-or-cgl=0.60.6
+  - coin-or-clp=1.17.7
+  - coin-or-osi=0.108.7
+  - coin-or-utils=2.11.6
+  - coincbc=2.10.8
+  - colorama=0.4.6
+  - comm=0.1.2
+  - configargparse=1.5.3
+  - connection_pool=0.0.3
+  - contourpy=1.0.7
+  - cryptography=39.0.0
+  - cvxpy=1.3.0
+  - cvxpy-base=1.3.0
+  - cycler=0.11.0
+  - dask-core=2023.1.0
+  - datrie=0.8.2
+  - dbus=1.13.6
+  - debugpy=1.6.6
+  - decorator=5.1.1
+  - defusedxml=0.7.1
+  - deprecation=2.1.0
+  - docutils=0.19
+  - dpath=2.1.4
+  - dropbox=11.36.0
+  - ecos=2.0.7.post1
+  - entrypoints=0.4
+  - exceptiongroup=1.1.0
+  - executing=1.2.0
+  - expat=2.5.0
+  - filechunkio=1.8
+  - filelock=3.9.0
+  - flit-core=3.8.0
+  - fontconfig=2.14.2
+  - fonttools=4.38.0
+  - freetype=2.12.1
+  - frozenlist=1.3.3
+  - fsspec=2023.1.0
+  - ftputil=5.0.4
+  - gettext=0.21.1
+  - gitdb=4.0.10
+  - gitpython=3.1.30
+  - glib=2.68.4
+  - glib-tools=2.68.4
+  - google-api-core=2.11.0
+  - google-api-python-client=2.74.0
+  - google-auth=2.16.0
+  - google-auth-httplib2=0.1.0
+  - google-cloud-core=2.3.2
+  - google-cloud-storage=2.7.0
+  - google-crc32c=1.1.2
+  - google-resumable-media=2.4.1
[…]
