### 2023-03-14 ed96bd9
benchmark/README: fix typo
