### 2023-03-14 f96dc33
update benchmark snakemake config and figure
