### 2023-03-14 #4823 @FabianHofmann
@torressa thanks for reviewing again. I've created a [README](https://github.com/PyPSA/linopy/tree/master/benchmark) for the benchmark which gives all necessary information how to reproduce the plots. Software and hardware specifications are now given [here](https://github.com/PyPSA/linopy/tree/master/benchmark#versions-specfications), these are also referenced in the paper.
As for the memory profiler, I have to admit that this is also a bit opaque for me. I have relied on the built-in memory profiler of [snakemake](https://snakemake.readthedocs.io/en/stable/). It seemed that sometimes it does not yield expected values. However with the current setting (fixed number of threads per job), it seems to be quite stable. Please let me know if you encounter problems.
I have adjusted the Gurobi reference, thanks for looking into it.

Let me know if there is further things to do :)
