### 2023-03-20 ac64741
fix representation of empty variables/linexpressions
