+++ paper/paper.md
@@ -40,7 +40,7 @@ The approach follows the idea that a variable $x(d_1, d_2, ..., d_K)$ may be def
 
 The array-based modelling approach does not only lead to more flexibility but also increases the overall performance. The following figure shows the [benchmark](https://github.com/PyPSA/linopy/tree/master/benchmark) against the AMLs JuMP, Pyomo, PuLP and CVXPY as well as the solver specific interface Gurobipy. The included AMLs packages are all open source and well-established and therefore suitable for comparison. Linopy outperforms all Python AMLs in memory efficiency and is close to CVXPY and Gurobipy in terms of speed while being faster than JuMP.
 
-![Benchmark of Linopy against comparable packages. The producing Snakemake workflow is available [here](https://github.com/PyPSA/linopy/tree/master/benchmark). The software and hardware specifications are detailed [here](https://github.com/PyPSA/linopy/tree/master/benchmark#versions-specfications). The benchmark is based on a 1-dimensional knapsack problem and uses the Gurobi solver. The overhead is calculated from the difference of the whole solving process via the AML and the solving process on the solver side alone. Note that the benchmark is hardly dependent on the complexity of the problem. Thus, adding more terms to the constraints, setting different kind of index labels or changing it to an purely linear problem does hardly have an effect on the overhead. \label{fig:benchmark}](figures/benchmark-overhead.pdf)
+![Benchmark of Linopy against comparable packages. The producing Snakemake workflow is available [here](https://github.com/PyPSA/linopy/tree/master/benchmark). The software and hardware specifications are detailed [here](https://github.com/PyPSA/linopy/tree/master/benchmark#versions-specfications). The benchmark is based on a 1-dimensional knapsack problem and uses the Gurobi solver. The overhead is calculated from the difference of the whole solving process via the AML and the solving process on the solver side alone. Note that the benchmark is hardly dependent on the complexity of the problem. Thus, adding more terms to the constraints, setting different kind of index labels or changing it to a purely linear problem does hardly have an effect on the overhead. \label{fig:benchmark}](figures/benchmark-overhead.pdf)
 
 Due to a strong alignment to the Xarray package, Linopy supports storing the optimization model as a NetCDF file [@rewNetCDFInterfaceScientific1990], which allows users to quickly share optimization problems with others. Using the [Paramiko package](https://paramiko.org), Linopy offers the user to send unsolved problems to a server and retrieve the solution after running the optimization remotely, which is particularly helpful if large computing resources are needed.
 
@@ -58,7 +58,7 @@ while other solvers such as `PIPS-IPM++` [@rehfeldtMassivelyParallelInteriorpoin
 
 # Related Research
 
-Linopy is used by several research projects and groups, mostly related to energy system modelling. The energy system modelling tool [PyPSA package](https://github.com/PyPSA/pypsa) [@brownPyPSAPythonPower2018], which is used by [various institutions](https://pypsa.readthedocs.io/en/latest/users.html) and builds the core of the [PyPSA-Eur workflow](https://github.com/PyPSA/pypsa-eur) [@horschPyPSAEurOpenOptimisation2018,@brownPyPSAPythonPower2018], uses Linopy as the primary optimization interface. The Fraunhofer Institute for Energy Economics and Energy System Technology is using Linopy in order to create an interface to GPU-based solvers. The German Aerospace Center uses Linopy for calculating stochastic optimization problems. Finally, a TU Berlin and Google Inc. cooperate on a [research project](https://github.com/PyPSA/247-cfe) that uses Linopy to analyze system-level impacts of 24/7 carbon-free electricity procurement in Europe.
[…]
