### 2023-03-20 #4823 @FabianHofmann
@Fei-Tao thank you for the instruction. I have published the new linopy version v0.1.4 (https://github.com/PyPSA/linopy/releases/tag/v0.1.4) which should be the associated with the paper. I have also created an archive of the software published in https://zenodo.org/record/7751989.

Please let me know if something is still missing
