### 2023-04-03 a737a2b
paper: update DOI's
