+++ paper/paper.bib
@@ -1,504 +1,545 @@
 @article{agrawalRewritingSystemConvex2018,
-  title   = {A Rewriting System for Convex Optimization Problems},
-  author  = {Agrawal, Akshay and Verschueren, Robin and Diamond, Steven and Boyd, Stephen},
-  year    = {2018},
+  title = {A Rewriting System for Convex Optimization Problems},
+  author = {Agrawal, Akshay and Verschueren, Robin and Diamond, Steven and Boyd, Stephen},
+  year = {2018},
+  month = jan,
   journal = {Journal of Control and Decision},
-  volume  = {5},
-  number  = {1},
-  pages   = {42--60}
+  volume = {5},
+  number = {1},
+  pages = {42--60},
+  issn = {2330-7706, 2330-7714},
+  doi = {10.1080/23307706.2017.1397554},
+  urldate = {2023-04-03},
+  langid = {english},
+  file = {/home/fabian/snap/zotero-snap/common/Zotero/storage/HWNIE65M/Agrawal et al. - 2018 - A rewriting system for convex optimization problem.pdf}
 }
 
 @misc{amploptimizationinc.AMPLStramlinedModeling,
-  title        = {{{AMPL}} - {{Stramlined}} Modeling for Real Optimization},
-  author       = {AMPL Optimization Inc.},
-  journal      = {AMPL},
-  abstract     = {Why AMPL? The AMPL system supports the entire optimization modeling lifecycle \textemdash{} formulation, testing, deployment, and maintenance \textemdash{} in an integrated way that promotes rapid development and reliable results. Using a high-level algebraic representation that describes optimization models in the...},
+  title = {{{AMPL}} - {{Stramlined}} Modeling for Real Optimization},
+  author = {AMPL Optimization Inc.},
+  journal = {AMPL},
+  urldate = {2022-07-22},
+  abstract = {Why AMPL? The AMPL system supports the entire optimization modeling lifecycle \textemdash{} formulation, testing, deployment, and maintenance \textemdash{} in an integrated way that promotes rapid development and reliable results. Using a high-level algebraic representation that describes optimization models in the...},
   howpublished = {https://ampl.com/},
-  langid       = {american},
-  file         = {/home/fabian/snap/zotero-snap/common/Zotero/storage/JNJKCUAT/ampl.com.html}
+  langid = {american},
+  file = {/home/fabian/snap/zotero-snap/common/Zotero/storage/JNJKCUAT/ampl.com.html}
 }
 
 @article{AnacondaSoftwareDistribution2020,
-  title     = {Anaconda Software Distribution},
-  year      = {2020},
-  journal   = {Anaconda Documentation},
+  title = {Anaconda Software Distribution},
+  year = {2020},
+  journal = {Anaconda Documentation},
   publisher = {{Anaconda Inc.}},
-  version   = {Vers. 2-2.4.0}
+  version = {Vers. 2-2.4.0}
 }
 
 @book{apsMOSEKOptimizerAPI2019,
-  title  = {{{MOSEK Optimizer API}} for {{Python}} 9.3.22},
-  author = {ApS, MOSEK},
-  year   = {2019}
+  title = {{{MOSEK Optimizer API}} for {{Python}} 9.3.22},
+  author = {ApS, {\relax MOSEK}},
+  year = {2019}
 }
 
 @article{bealGEKKOOptimizationSuite2018,
-  title    = {{{GEKKO Optimization Suite}}},
-  author   = {Beal, Logan and Hill, Daniel and Martin, R. and Hedengren, John},
-  year     = {2018},
-  month    = jul,
-  journal  = {Processes},
-  volume   = {6},
-  number   = {8},
-  pages    = {106},
-  issn     = {2227-9717},
[…]
