+++ paper/paper.bib
@@ -1,19 +1,3 @@
-@article{agrawalRewritingSystemConvex2018,
-  title = {A Rewriting System for Convex Optimization Problems},
-  author = {Agrawal, Akshay and Verschueren, Robin and Diamond, Steven and Boyd, Stephen},
-  year = {2018},
-  month = jan,
-  journal = {Journal of Control and Decision},
-  volume = {5},
-  number = {1},
-  pages = {42--60},
-  issn = {2330-7706, 2330-7714},
-  doi = {10.1080/23307706.2017.1397554},
-  urldate = {2023-04-03},
-  langid = {english},
-  file = {/home/fabian/snap/zotero-snap/common/Zotero/storage/HWNIE65M/Agrawal et al. - 2018 - A rewriting system for convex optimization problem.pdf}
-}
-
 @misc{amploptimizationinc.AMPLStramlinedModeling,
   title = {{{AMPL}} - {{Stramlined}} Modeling for Real Optimization},
   author = {AMPL Optimization Inc.},
@@ -131,7 +115,7 @@
   publisher = {{Springer US}},
   address = {{Boston, MA}},
   doi = {10.1007/978-1-4613-0215-5_8},
-  urldate = {2022-08-29},
+  urldate = {2023-04-03},
   isbn = {978-1-4613-7945-4 978-1-4613-0215-5}
 }
 
@@ -154,18 +138,6 @@
   urldate = {2022-09-02},
   langid = {english},
   file = {/home/fabian/snap/zotero-snap/common/Zotero/storage/SLY4DL3G/Dantzig - 1963 - Linear Programming and Extensions.pdf}
-}
-
-@book{dantzigLinearProgrammingExtensions1998,
-  title = {Linear Programming and Extensions},
-  author = {Dantzig, George Bernard},
-  year = {1998},
-  series = {Princeton Landmarks in Mathematics and Physics},
-  publisher = {{Princeton University Press}},
-  address = {{Princeton (N. J.)}},
-  isbn = {978-0-691-08000-0 978-0-691-05913-6},
-  langid = {english},
-  lccn = {519.7}
 }
 
 @manual{daskdevelopmentteamDaskLibraryDynamic2016,
@@ -239,14 +211,6 @@
   urldate = {2022-08-16},
   abstract = {Practical large-scale mathematical programming involves more than just the application of an algorithm to minimize or maximize an objective function. Before any optimizing routine can be invoked, considerable effort must be expended to formulate the underlying model and to generate the requisite computational data structures. AMPL is a new language designed to make these steps easier and less error-prone. AMPL closely resembles the symbolic algebraic notation that many modelers use to describe mathematical programs, yet it is regular and formal enough to be processed by a computer system; it is particularly notable for the generality of its syntax and for the variety of its indexing operations. We have implemented an efficient translator that takes as input a linear AMPL model and associated data, and produces output suitable for standard linear programming optimizers. Both the language and the translator admit straightforward extensions to more general mathematical programs that incorporate nonlinear expressions or discrete variables.},
   langid = {english}
-}
-
-@misc{gamsdevelopmentcorporationGeneralAlgebraicModeling,
-  title = {General {{Algebraic Modeling System}} ({{GAMS}})},
-  author = {GAMS Development Corporation},
-  urldate = {2022-07-22},
-  howpublished = {https://www.gams.com/},
-  file = {/home/fabian/snap/zotero-snap/common/Zotero/storage/LCJFIAG2/www.gams.com.html}
 }
 
 @misc{GLPKGNUProject,
