### 2023-04-03 #4823 @Fei-Tao
@FabianHofmann, could you make a new release of this software that includes the changes resulting from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the Zenodo/figshare archive, please make sure that:

The title of the archive is the same as the JOSS paper title
That the authors of the archive are the same as the JOSS paper authors
