### 2023-04-03 #4823 @FabianHofmann
> @Fei-Tao thank you for the instruction. I have published the new linopy version v0.1.4 (https://github.com/PyPSA/linopy/releases/tag/v0.1.4) which should be the associated with the paper. I have also created an archive of the software published in https://zenodo.org/record/7751989.

Hey @Fei-Tao this was already done some days ago, see above. Does that meet all requirements?
