### 2023-04-21 #4823 @FabianHofmann
@Fei-Tao thanks for the reply and sorry if I bother you again. I actually think the `recommend-accept` command did not work properly. Normally, the issue should have a `recommend-accept` tag and a message created from the bot, like [here](https://github.com/openjournals/joss-reviews/issues/4149#issuecomment-1513151793). Typically, the JOSS team is very responsive and an answer after the recommendation takes a couple of days at max.
