### 2023-04-21 7f993f6
model: reindex constraint labels by model.constraints if necessary
