### 2023-04-21 e1925fc
Upping the float precision in lp files for #111, remove tailing zeros for #112
