### 2023-04-21 ae54e0a
add test for reindexing constraint labels
