+++ test/test_variables.py
@@ -45,6 +45,15 @@ def test_variables_assignment_with_merge():
     assert m.variables.labels.var0[-1].item() != -1
 
 
+def test_variables_assignment_with_reindex(m):
+    shuffled_coords = [pd.Index([2, 1, 3, 4, 6, 5, 7, 9, 8, 0], name="first")]
+    with pytest.warns(UserWarning):
+        a = m.add_variables(coords=shuffled_coords, name="a")
+    assert a.indexes["first"].equals(m["x"].indexes["first"])
+    # check if labels are monotonically increasing
+    assert np.all(np.diff(np.ravel(a.labels)) > 0)
+
+
 def test_scalar_variables_name_counter():
     m = Model()
     m.add_variables()
